% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r, R/sql_translate_env.R
\name{Athena}
\alias{Athena}
\alias{db_data_type.AthenaConnection}
\alias{sql_translate_env.AthenaConnection}
\title{AthenaJDBC}
\usage{
Athena(identifier.quote = "`")

db_data_type.AthenaConnection(con, fields, ...)

sql_translate_env.AthenaConnection(con)
}
\description{
AthenaJDBC
}
\keyword{internal}
