% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{class}
\name{AthenaConnection-class}
\alias{AthenaConnection-class}
\title{AthenaJDBC}
\arguments{
\item{jc}{job ref}

\item{identifier.quote}{how to quote identifiers}

\item{fetch_size}{Athena results fetch size}
}
\description{
AthenaJDBC
}
