% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.colors.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{srn.colors}
\format{An object of class \code{character} of length 16.}
\usage{
cbPalette
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for SRN modeling
}
\examples{
srn.colors()
pie(rep(1,length(colorsX_Basic)),label=names(colorsX_Basic),col=colorsX_Basic)
}
\keyword{colors,}
\keyword{palette}
