% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEncodedData.R
\name{createDataForEncoder}
\alias{createDataForEncoder}
\title{Create propensity scores}
\usage{
createDataForEncoder(
  cohortMethodData,
  population = NULL,
  mapping = NULL,
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  maxCohortSizeForFitting = 250000,
  fileName = NULL,
  weight = "default",
  tidyCovariate = T
)
}
\arguments{
\item{cohortMethodData}{An object of type [CohortMethodData] as generated using
[getDbCohortMethodData()].}

\item{population}{A data frame describing the population. This should at least have a
`rowId` column corresponding to the `rowId` column in the
[CohortMethodData] covariates object and a `treatment` column.
If population is not specified, the full population in the
[CohortMethodData] will be used.}

\item{excludeCovariateIds}{Exclude these covariates from the propensity model.}

\item{includeCovariateIds}{Include only these covariates in the propensity model.}

\item{maxCohortSizeForFitting}{If the target or comparator cohort are larger than this number, they
will be downsampled before fitting the propensity model. The model
will be used to compute propensity scores for all subjects. The
purpose of the sampling is to gain speed. Setting this number to 0
means no downsampling will be applied.}

\item{removeRedundancy}{If true, the function will remove the redundant covariates by using
`FeatureExtraction::tidyCovariateData` function. Default setting is true.}
}
\description{
Creates propensity scores using a regularized logistic regression.
}
\examples{
data(cohortMethodDataSimulationProfile)
cohortMethodData <- simulateCohortMethodData(cohortMethodDataSimulationProfile, n = 1000)
ps <- createPs(cohortMethodData)

}
