% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetisCohortMethod.R
\name{createReduceDimArgs}
\alias{createReduceDimArgs}
\title{Create a parameter object for the function reduceDim}
\usage{
createReduceDimArgs(pathToAutoencoderWeight = "", pathToMap = "")
}
\arguments{
\item{pathToMap}{Restrict the analysis to the period when both exposures are
observed?}

\item{pathToEncoder}{Should only the first exposure per subject be included?}
}
\description{
Create a parameter object for the function reduceDim
}
\details{
Create an object defining the parameter values.
}
