% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_A23.globaltech_retirement}
\alias{data_A23.globaltech_retirement}
\title{data_A23.globaltech_retirement}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/investCalcs/A23.globaltech_retirement.csv",sep="")
}
\usage{
data_A23.globaltech_retirement
}
\description{
data_A23.globaltech_retirement
}
\examples{
\dontrun{
 library(metis);
 metis::data_A23.globaltech_retirement
}
}
\keyword{datasets}
