% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cap_cost_int_cool}
\alias{data_cap_cost_int_cool}
\title{data_cap_cost_int_cool}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/investCalcs/L2233.GlobalIntTechCapital_elec_cool.csv", sep="")
}
\usage{
data_cap_cost_int_cool
}
\description{
data_cap_cost_int_cool
}
\examples{
\dontrun{
 library(metis);
 metis::data_cap_cost_int_cool
}
}
\keyword{datasets}
