% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cap_cost_tech}
\alias{data_cap_cost_tech}
\title{data_cap_cost_tech}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/investCalcs/L2233.GlobalTechCapital_elecPassthru.csv", sep="")
}
\usage{
data_cap_cost_tech
}
\description{
data_cap_cost_tech
}
\examples{
\dontrun{
 library(metis);
 metis::data_cap_cost_tech
}
}
\keyword{datasets}
