% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_capac_fac_int}
\alias{data_capac_fac_int}
\title{data_capac_fac_int}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/investCalcs/L223.GlobalIntTechCapFac_elec.csv", sep="")
}
\usage{
data_capac_fac_int
}
\description{
data_capac_fac_int
}
\examples{
\dontrun{
 library(metis);
 metis::data_capac_fac_int
}
}
\keyword{datasets}
