% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_capfactors}
\alias{data_capfactors}
\title{data_capfactors}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/capacity_factor_by_elec_gen_subsector.csv",sep="")
}
\usage{
data_capfactors
}
\description{
data_capfactors
}
\examples{
\dontrun{
 library(metis);
 metis::data_capfactors
}
}
\keyword{datasets}
