% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tech_mapping}
\alias{data_tech_mapping}
\title{data_tech_mapping}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/investCalcs/agg_tech_mapping.csv", sep="")
}
\usage{
data_tech_mapping
}
\description{
data_tech_mapping
}
\examples{
\dontrun{
 library(metis);
 metis::tech_mapping
}
}
\keyword{datasets}
