% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-dbGetQuery.R
\docType{methods}
\name{dbGetInfo,AthenaDriver-method}
\alias{dbGetInfo,AthenaDriver-method}
\title{Retrieve connection/driver/database metadata}
\usage{
\S4method{dbGetInfo}{AthenaDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{driver/connection}

\item{...}{unused}
}
\description{
Retrieve connection/driver/database metadata
}
\keyword{internal}
