% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbListTables,AthenaConnection-method}
\alias{dbListTables,AthenaConnection-method}
\title{AthenaJDBC}
\usage{
\S4method{dbListTables}{AthenaConnection}(conn, pattern = "*", schema,
  ...)
}
\arguments{
\item{conn}{Athena connection}

\item{pattern}{table name pattern}

\item{schema}{Athena schema name}

\item{...}{unused}
}
\description{
AthenaJDBC
}
