% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleGCAMproj}
\alias{exampleGCAMproj}
\title{Example GCAM .proj file}
\format{
.proj saved as a list
}
\source{
GCAM run proj file.
}
\usage{
exampleGCAMproj
}
\description{
Example GCAM .proj file
}
\examples{
\dontrun{
 library(metis); library(rgcam)
 dataProjLoaded <- metis::exampleGCAMproj
 rgcam::listScenarios(dataProjLoaded)  # List of Scenarios in GCAM database
 rgcam::listQueries(dataProjLoaded) # List of Queries in queryxml
}
}
\keyword{datasets}
