% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMapDataParam}
\alias{exampleMapDataParam}
\title{Example data by param to plot in Maps}
\format{
R table or .csv
}
\source{
metis.readgcam() run outputs saved
}
\usage{
exampleMapDataParam
}
\description{
Example data by param to plot in Maps
}
\examples{
\dontrun{
 library(metis);
 mapData <- metis::exampleMapDataParam
}
}
\keyword{datasets}
