% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{generateSingleSiteResult}
\alias{generateSingleSiteResult}
\title{Generation cmObject and fitted PS model}
\usage{
generateSingleSiteResult(connectionDetails, oracleTempSchema = NULL,
  outputFolder, databaseId = "Unknown", databaseName = "Unknown",
  databaseDescription = "Unknown", cdmDatabaseSchema,
  outcomeDatabaseSchema = cdmDatabaseSchema, outcomeTable = "outcome",
  nestingDatabaseSchema = cdmDatabaseSchema, nestingTable = "nesting",
  createCohorts = TRUE, createCmObject = TRUE, fitPsModel = TRUE,
  uploadResult = FALSE, maxCores = 4)
}
\description{
Generation cmObject and fitted PS model
}
