% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid025}
\alias{grid025}
\title{Empty global grid 0.25 degrees}
\format{R tibble}
\source{
JGCRI
}
\usage{
grid025
}
\description{
Empty global grid 0.25 degrees
}
\examples{
\dontrun{
library(metis)
head(grid025)
}
}
\keyword{datasets}
