% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid050}
\alias{grid050}
\title{US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.}
\format{R tibble}
\source{
JGCRI
}
\usage{
grid050
}
\description{
US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.
}
\examples{
\dontrun{
library(metis)
head(grid050)
}
}
\keyword{datasets}
