% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_pop_GPWv4To2015}
\alias{grid_pop_GPWv4To2015}
\title{Gridded Population of the World, Version 4 (GPWv4)
https://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-count-rev11/data-download#}
\format{
R tibble
}
\source{
SEDAC, https://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-count-rev11/data-download#
}
\usage{
grid_pop_GPWv4To2015
}
\description{
Gridded Population of the World, Version 4 (GPWv4)
https://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-count-rev11/data-download#
}
\examples{
\dontrun{
library(metis)
head(data_pop_GPWv4To2015)
}
}
\keyword{datasets}
