% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapCountries}
\alias{mapCountries}
\title{World Map of Countries}
\format{A SpatialPolygonsDataFrame}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapCountries
}
\description{
World Map of Countries
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapCountries)
 head(mapCountries@data)
}
}
\keyword{datasets}
