% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasins}
\alias{mapGCAMBasins}
\title{GCAM Basins}
\format{
A SpatialPolygonsDataFrame
}
\source{
From JGCRI confluence page. \url{https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files}
}
\usage{
mapGCAMBasins
}
\description{
GCAM Basins
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMBasins)
 head(mapGCAMBasins@data)
}
}
\keyword{datasets}
