% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLand}
\alias{mapGCAMLand}
\title{GCAM Land}
\format{
A SpatialPolygonsDataFrame
}
\source{
From the moirai project.
Shapefile received directly from Alan V. Di Vittorio.
Component files available at:
\url{https://github.com/JGCRI/moirai/blob/master/ancillary/moirai_valid_boundaries.zip}
\url{https://github.com/JGCRI/moirai/blob/build-r-package/tests/testthat/test_compare_raster_w_outputs.R}
}
\usage{
mapGCAMLand
}
\description{
GCAM Land
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMLand)
 head(mapGCAMLand@data)
}
}
\keyword{datasets}
