% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLandUS49}
\alias{mapGCAMLandUS49}
\title{GCAM Land cropped to US49}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from metis maps.
}
\usage{
mapGCAMLandUS49
}
\description{
GCAM Land cropped to US49
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMLandUS49)
 head(mapGCAMLandUS49@data)
}
}
\keyword{datasets}
