% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLandUS52}
\alias{mapGCAMLandUS52}
\title{GCAM Land cropped to US52}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from metis maps
}
\usage{
mapGCAMLandUS52
}
\description{
GCAM Land cropped to US52
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMLandUS52)
 head(mapGCAMLandUS52@data)
}
}
\keyword{datasets}
