% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32}
\alias{mapGCAMReg32}
\title{GCAM 32 Regions}
\format{A SpatialPolygonsDataFrame}
\source{
From JGCRI confluence page. \url{https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files}
}
\usage{
mapGCAMReg32
}
\description{
GCAM 32 Regions
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMReg32)
 head(mapGCAMReg32@data)
}
}
\keyword{datasets}
