% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32US52}
\alias{mapGCAMReg32US52}
\title{Merge of GCAM 32 and US52}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMReg32US52
}
\description{
Merge of GCAM 32 and US52
}
\examples{
\dontrun{
library(metis)
head(mapGCAMReg32US52)
}
}
\keyword{datasets}
