% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapHydroShed1}
\alias{mapHydroShed1}
\title{HydroSHEDS level 1}
\format{
A SpatialPolygonsDataFrame
}
\source{
Lehner, B., Grill G. (2013): Global river hydrography and network routing:
baseline data and new approaches to study the world’s large river systems.
Hydrological Processes, 27(15): 2171–2186. \url{https://www.hydrosheds.org/page/hydrobasins}
}
\usage{
mapHydroShed1
}
\description{
HydroSHEDS level 1
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapHydroShed1)
 head(mapHydroShed1@data)
}
}
\keyword{datasets}
