% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapHydroShed4}
\alias{mapHydroShed4}
\title{HydroSHEDS level 4}
\format{A SpatialPolygonsDataFrame}
\source{
Lehner, B., Grill G. (2013): Global river hydrography and network routing:
baseline data and new approaches to study the world’s large river systems.
Hydrological Processes, 27(15): 2171–2186. \url{https://www.hydrosheds.org/page/hydrobasins}
}
\usage{
mapHydroShed4
}
\description{
HydroSHEDS level 4
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapHydroShed4)
 head(mapHydroShed4@data)
}
}
\keyword{datasets}
