% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasin32Reg}
\alias{mapIntersectGCAMBasin32Reg}
\title{Intersection of GCAM Basins and GCAM 32 Regions
Intersection of GCAM Basins and GCAM 32 Regions.}
\format{R tibble}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasin32Reg
}
\description{
Intersection of GCAM Basins and GCAM 32 Regions
Intersection of GCAM Basins and GCAM 32 Regions.
}
\examples{
\dontrun{
library(metis)
head(mapIntersectGCAMBasin32Reg)
}
}
\keyword{datasets}
