% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasinCountry}
\alias{mapIntersectGCAMBasinCountry}
\title{Intersection of GCAM Basins and countries.}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasinCountry
}
\description{
Intersection of GCAM Basins and countries.
}
\examples{
\dontrun{
library(metis)
head(mapIntersectGCAMBasinCountry)
}
}
\keyword{datasets}
