% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasinUS52}
\alias{mapIntersectGCAMBasinUS52}
\title{Intersection of GCAM Basins and US 52 Regions.}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasinUS52
}
\description{
Intersection of GCAM Basins and US 52 Regions.
}
\examples{
\dontrun{
library(metis)
head(mapIntersectGCAMBasinUS52)
}
}
\keyword{datasets}
