% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapStates}
\alias{mapStates}
\title{World Map of States}
\format{
A SpatialPolygonsDataFrame
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapStates
}
\description{
World Map of States
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapStates)
 head(mapStates@data)
}
}
\keyword{datasets}
