% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS49County}
\alias{mapUS49County}
\title{US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.}
\format{A SpatialPolygonsDataFrame}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapUS49County
}
\description{
US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapUS49County)
 head(mapUS49County@data)
}
}
\keyword{datasets}
