% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS49HUC2}
\alias{mapUS49HUC2}
\title{USGS Hydrological Unit Code (HUC)}
\format{A SpatialPolygonsDataFrame}
\source{
\url{https://water.usgs.gov/GIS/huc.html} \url{https://nrcs.app.box.com/v/huc}
}
\usage{
mapUS49HUC2
}
\description{
USGS Hydrological Unit Code (HUC)
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapUS49HUC2)
 head(mapUS49HUC2@data)
}
}
\keyword{datasets}
