% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52}
\alias{mapUS52}
\title{US 52 States
Includes Alaska, Hawaii and Puerto Rico as well as DC.}
\format{
A SpatialPolygonsDataFrame
}
\source{
US Census bureau. \url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
mapUS52
}
\description{
US 52 States
Includes Alaska, Hawaii and Puerto Rico as well as DC.
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapUS52)
 head(mapUS52@data)
}
}
\keyword{datasets}
