% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52Compact}
\alias{mapUS52Compact}
\title{US 52 States Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)}
\format{
A SpatialPolygonsDataFrame
}
\source{
US Census bureau. \url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
mapUS52Compact
}
\description{
US 52 States Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapUS52Compact)
 head(mapUS52Compact@data)
}
}
\keyword{datasets}
