% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.irio.R
\name{met.irio}
\alias{met.irio}
\title{metis.irio}
\usage{
met.irio(Z0 = NULL, D0 = NULL, X0 = NULL, D = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""))
}
\arguments{
\item{Z0}{Default = NULL,}

\item{D0}{Default = NULL,}

\item{X0}{Default = NULL,}

\item{D}{Default = NULL,}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
