% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mrio.R
\name{met.mrio}
\alias{met.mrio}
\title{metis.mrio}
\usage{
met.mrio(Z0 = NULL, Q0 = NULL, D0 = NULL, X0 = NULL, D = NULL,
  n_regions = 2, dirOutputs = paste(getwd(), "/outputs", sep = ""))
}
\arguments{
\item{Z0}{Initial intermediate flow matrix. All diagnol matrices 0. Default = NULL,}

\item{Q0}{Initial trade matrix. Columns sum to 100. Default = NULL,}

\item{D0}{Initital External demand. Default = NULL,}

\item{X0}{Initial total Demand internal and external. Default = NULL,}

\item{D}{External demand or Household demand. Default = NULL,}

\item{n_regions}{Number of regions. Default = NULL,}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
