% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.bia.R
\name{metis.bia}
\alias{metis.bia}
\title{metis.bia}
\usage{
metis.bia(
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  regionsSelect = NULL,
  regionsSelectDiagnostic = NULL,
  dataProjFile = "dataProj.proj",
  gcamdatabase = NULL,
  gridChoice = "grid_050",
  diagnosticsON = F,
  subsectorNAdistribute = "even",
  folderName = NULL,
  nameAppend = ""
)
}
\arguments{
\item{dirOutputs}{Default=paste(getwd(),"/outputs",sep  Default=""). Location for outputs.}

\item{regionsSelect}{The regions to analyze in a vector. Example c('Colombia','Argentina')}

\item{regionsSelectDiagnostic}{Default = NULL (sets to regionsSelect). Choose the regions to run diagnostics.}

\item{dataProjFile}{Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.}

\item{gcamdatabase}{Default NULL. Full path to gcamdatabase.}

\item{gridChoice}{Default = "grid_050" . Choice of whether to use 50 km x 50 km grid cells ("grid_050") or 25 km x 25 km ("grid_025").}

\item{diagnosticsON}{Default = T.}

\item{subsectorNAdistribute}{Default = "even". Choose "even" for even distribution or "totalOther" to distribute based on sum of all other subsectors..}

\item{folderName}{Default = NULL,}

\item{nameAppend}{Default=""}
}
\value{
A tibble with GCAM electricity generation distributed on a grid for selected regions
}
\description{
This function downscales GCAM electricity generation and installed capacity onto a grid, based on WRI PowerWatch dataset of present capacity
}
\keyword{downscale,}
\keyword{downscaled}
\keyword{downscaling,}
\keyword{electricity,}
\keyword{gcam,}
\keyword{generation,}
\keyword{gridded,}
