% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.boundaries.R
\name{metis.boundaries}
\alias{metis.boundaries}
\title{metis.boundaries}
\usage{
metis.boundaries(boundaryRegShape = NULL, boundaryRegShpFolder = NULL,
  boundaryRegShpFile = NULL, boundaryRegCol = NULL,
  boundaryRegionsSelect = NULL, subRegShape = NULL,
  subRegShpFolder = NULL, subRegShpFile = NULL, subRegCol = NULL,
  subRegionsSelect = NULL, subRegType = "subRegType",
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "",
  expandPercent = 2, overlapShape = NULL, overlapShpFolder = NULL,
  overlapShpFile = NULL, labelsSize = 1.2, fillcolorNA = NULL,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  extendedFillColor = "grey75", extendedBGColor = "lightblue1",
  extendedHighLightColor = "cornsilk1", extendedLabelsColor = "grey30",
  extdendedLabelSize = 0.7, extension = T, fillPalette = "Spectral",
  cropSubShape2Bound = T, grids = NULL)
}
\arguments{
\item{boundaryRegShape}{Default=NULL,}

\item{boundaryRegShpFolder}{Default= NULL . Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default=""),}

\item{boundaryRegShpFile}{Default=NULL . Suggested paste("ne_10m_admin_0_countries",sep  Default=""),}

\item{boundaryRegCol}{Default=NULL. Suggested "NAME_0",}

\item{boundaryRegionsSelect}{Default=NULL. Suggested GADM country Names}

\item{subRegShape}{Default=NULL,}

\item{subRegShpFolder}{Default=NULL. Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default=""),}

\item{subRegShpFile}{Default=NULL. Suggested paste("ne_10m_admin_1_states_provinces",sep  Default=""),}

\item{subRegCol}{Default="NULL Suggested for states "name",}

\item{subRegionsSelect}{Default=NULL,}

\item{subRegType}{Default="subRegType",}

\item{dirOutputs}{Default=paste(getwd(),"/outputs",sep  Default=""),}

\item{nameAppend}{Default="",}

\item{expandPercent}{Default=2,}

\item{overlapShape}{Default = NULL}

\item{overlapShpFolder}{Default = NULL. For Basins use paste(getwd(),"/dataFiles/gis/basin_gcam",sep=""),}

\item{overlapShpFile}{Default = NULL. For Basins use ="Global235_CLM_final_5arcmin_multipart"}

\item{labelsSize}{Default =1.2}

\item{fillcolorNA}{Default =NULL}

\item{projX}{Default ="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",}

\item{extendedFillColor}{Default = "grey75",}

\item{extendedBGColor}{Default = "lightblue1",}

\item{extendedHighLightColor}{Default = "cornsilk1",}

\item{extendedLabelsColor}{Default = "grey30"}

\item{extdendedLabelSize}{Default =0.7}

\item{extension}{Default = T}

\item{fillPalette}{Default ="Spectral",}

\item{cropSubShape2Bound}{Default = T. Set to False if subregion shape is larger than boundary, but desired fro extension.}

\item{grids}{Default = NULL. Suggested is c(paste(getwd(),"/dataFiles/grids/emptyGrids/grid_025.csv",sep=""),
paste(getwd(),"/dataFiles/grids/emptyGrids/grid_050.csv",sep=""))
This may happen in the case of disputed boundaries.}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function takes a .csv file with gridded lat, long data and aggregates
the data by spatial boundaries given different shapefiles.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
