% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.downscale.R
\name{metis.downscale}
\alias{metis.downscale}
\title{metis.downscale}
\usage{
metis.downscale(
  polygonTable = NULL,
  gridTable = NULL,
  downscaleRes = "grid050",
  shape = NULL,
  shapeFolder = NULL,
  shapeFile = NULL,
  colName = "subRegion",
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  fname = "downscale",
  nameAppend = NULL,
  folderName = NULL,
  saveFile = T,
  printFig = F
)
}
\arguments{
\item{polygonTable}{Default = NULL,}

\item{gridTable}{Default = NULL. Full path to grid file or R tibble or for pre-loaded data one of "pop".}

\item{downscaleRes}{Default = grid050. Can Choose grid025 or provide a dataframe with equally spaced lat and lon.}

\item{shape}{Default = NULL,}

\item{shapeFolder}{Default = NULL,}

\item{shapeFile}{Default = NULL,}

\item{colName}{Default = "subRegion",}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{fname}{Default = "gridByPoly"}

\item{nameAppend}{Default = NULL}

\item{folderName}{Default = NULL}

\item{saveFile}{Default = F. If want csv output then change to T}

\item{printFig}{Default = F. If grid overlap with shape is wanted}
}
\value{
Prints out graphic
}
\description{
This function downscales data by subregion by given gridded data
}
\keyword{grid,}
\keyword{polygon}
\keyword{shape,}
