% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.grid2poly.R
\name{metis.grid2poly}
\alias{metis.grid2poly}
\title{metis.grid2poly}
\usage{
metis.grid2poly(grid = NULL, boundaryRegionsSelect = NULL,
  subRegShape = NULL, subRegShpFolder = NULL, subRegShpFile = NULL,
  subRegCol = NULL, subRegType = "subRegType", aggType = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "",
  labelsSize = 1.2)
}
\arguments{
\item{grid}{Default=NULL,}

\item{boundaryRegionsSelect}{Default=NULL,}

\item{subRegShape}{Default=NULL,}

\item{subRegShpFolder}{Default = NULL, Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default="")}

\item{subRegShpFile}{Default = NULL, paste("ne_10m_admin_1_states_provinces",sep  Default=""),}

\item{subRegCol}{Default = NULL, "NAME_1",}

\item{subRegType}{Default="subRegType",}

\item{aggType}{Default=NULL,}

\item{dirOutputs}{Default=paste(getwd(),"/outputs",sep  Default=""),}

\item{nameAppend}{Default="",}

\item{labelsSize}{Default =1.2}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function takes a .csv file with gridded lat, long data and aggregates
the data by spatial boundaries given different shapefiles.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
