% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.gridByPoly.R
\name{metis.gridByPoly}
\alias{metis.gridByPoly}
\title{metis.gridByPoly}
\usage{
metis.gridByPoly(
  gridTable = NULL,
  shape = NULL,
  shapeFolder = NULL,
  shapeFile = NULL,
  colName = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  fname = "gridByPoly",
  folderName = "folderNameDefault",
  saveFile = F
)
}
\arguments{
\item{gridTable}{Default = NULL. Full path to grid file.}

\item{shape}{Default = NULL,}

\item{shapeFolder}{Default = NULL,}

\item{shapeFile}{Default = NULL,}

\item{colName}{Default = NULL,}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{fname}{Default = "gridByPoly"}

\item{folderName}{Default ="folderNameDefault",}

\item{saveFile}{Default = F. If want csv output then change to T}
}
\value{
Prints out graphic
}
\description{
This function finds the grids located within a given shapefiles regions
}
\keyword{grid,}
\keyword{polygon}
\keyword{shape,}
