% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.map.R
\name{metis.map}
\alias{metis.map}
\title{metis.map}
\usage{
metis.map(dataPolygon = NULL, dataGrid = NULL, dataRaster = NULL,
  shpFolder = paste(getwd(), "/dataFiles/gis/admin_gadm36", sep = ""),
  shpFile = paste("gadm36_1", sep = ""), shpName = "NAME_0",
  fillPalette = "Spectral", borderColor = "gray20", lwd = 1,
  lty = 1, bgColor = "white", frameShow = F, fillColumn = NULL,
  labels = F, labelsSize = 1.2, labelsColor = "black",
  labelsAutoPlace = F, figWidth = 9, figHeight = 7,
  legendWidth = -1, legendShow = F, legendOutside = T,
  legendTextSize = 1, legendTitleSize = 2,
  legendOutsidePosition = NULL, legendPosition = NULL,
  legendDigits = NULL, legendTitle = "Legend",
  legendStyle = "pretty", legendFixedBreaks = 5, legendBreaks = NULL,
  pdfpng = "png", underLayer = NULL, overLayer = NULL,
  printFig = T, fileName = "map", dirOutputs = paste(getwd(),
  "/outputs", sep = ""), facetFreeScale = F, facetRows = NA,
  facetCols = 3, facetLabelColor = "grey75", facetLabelSize = 1.5,
  alpha = 1, rasterCoverNegShape = F)
}
\arguments{
\item{dataPolygon}{Default = NULL,}

\item{dataGrid}{Default = NULL,}

\item{dataRaster}{Default = NULL,}

\item{shpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36_1",sep Default = ""),}

\item{shpFile}{Default = paste("gadm36_1",sep Default = ""),}

\item{shpName}{Default = "NAME_0",}

\item{fillPalette}{Default = "Spectral",}

\item{borderColor}{Default = "gray20",}

\item{lwd}{Default = 1,}

\item{lty}{Default = 1,}

\item{bgColor}{Default = "white",}

\item{frameShow}{Default = F,}

\item{fillColumn}{Default = NULL, # Or give column data with}

\item{labels}{Default = F,}

\item{labelsSize}{Default = 1.2,}

\item{labelsColor}{Default = "black",}

\item{labelsAutoPlace}{Default = F,}

\item{figWidth}{Default = 9,}

\item{figHeight}{Default = 7,}

\item{legendWidth}{Default = -1,}

\item{legendShow}{Default = F,}

\item{legendOutside}{Default = T,}

\item{legendTextSize}{Default = 0.8,}

\item{legendTitleSize}{Default = 1,}

\item{legendOutsidePosition}{Default = NULL, # "right","left","top","bottom", "center"}

\item{legendPosition}{Default = NULL, # c("RIGHT','top') - RIGHT LEFT TOP BOTTOM}

\item{legendDigits}{Default = NULL,}

\item{legendTitle}{Default = "Legend",}

\item{legendStyle}{Default = "pretty",}

\item{legendFixedBreaks}{Default = "5",}

\item{legendBreaks}{Default = NULL,}

\item{pdfpng}{Default = "png",}

\item{underLayer}{Default = NULL,}

\item{overLayer}{Default = NULL,}

\item{printFig}{Default = T,}

\item{fileName}{Default = "map",}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep Default = ""),}

\item{facetFreeScale}{Default = F,}

\item{facetRows}{Default = NA,}

\item{facetCols}{Default = 3,}

\item{facetLabelColor}{Default = "grey75",}

\item{facetLabelSize}{Default = 1.5,}

\item{alpha}{Default = 1}

\item{rasterCoverNegShape}{Default =T}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the metis package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
