% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mapFind.R
\name{metis.mapFind}
\alias{metis.mapFind}
\title{metis.mapFind}
\usage{
metis.mapFind(dataTbl)
}
\arguments{
\item{dataTbl}{Palette name to view the palette colors. Eg. metis.colors("pal_Basic")}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded metis maps.
}
\examples{
library(metis)
data = data.frame(subRegion=c("FL","ID","MO"),value=c(-2,3,14))
metis::metis.mapFind(data)
}
\keyword{find}
\keyword{map,}
