% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mapsProcess.R
\name{metis.mapsProcess}
\alias{metis.mapsProcess}
\title{metis.mapsProcess}
\usage{
metis.mapsProcess(
  polygonDataTables = NULL,
  gridDataTables = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  folderName = "",
  xRange = "All",
  labels = F,
  labelsSize = 1,
  subRegShape = NULL,
  subRegShpFolder = NULL,
  subRegShpFile = NULL,
  subRegCol = "subRegion",
  dirNameAppend = "",
  nameAppend = "",
  legendOutsideSingle = T,
  legendOutsidePosition = NULL,
  legendPosition = NULL,
  legendFixedBreaks = 5,
  legendTitleSizeO = 1,
  legendTextSizeO = 0.6,
  legendTitleSizeI = 1,
  legendTextSizeI = 0.6,
  animateOn = T,
  fps = 1,
  scenRef = NULL,
  scenDiff = NULL,
  extension = F,
  boundaryRegShape = NULL,
  boundaryRegShpFolder = NULL,
  boundaryRegShpFile = NULL,
  boundaryRegCol = "subRegion",
  boundaryRegionsSelect = NULL,
  extendedLabels = T,
  extendedFillColor = "grey75",
  extendedBGColor = "lightblue1",
  extendedHighLightColor = "cornsilk1",
  extendedLabelsColor = "grey30",
  extdendedLabelSize = 0.7,
  extendedShape = NULL,
  extendedShapeCol = NULL,
  expandPercent = 3,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  figWidth = 6,
  figHeight = 7,
  scaleRange = NULL,
  scaleRangeDiffAbs = NULL,
  scaleRangeDiffPrcnt = NULL,
  paramsSelect = "All",
  multifacetsOn = F,
  multiFacetCols = "multiFacetCol",
  multiFacetRows = "multiFacetRow",
  legendOutsideMulti = T,
  legendPositionMulti = NULL,
  legendTitleSizeMulti = NULL,
  legendTextSizeAnim = NULL,
  legendTextSizeMulti = NULL,
  refMultiA = NULL,
  refMultiB = NULL,
  chosenRefMeanYears = NULL,
  mapTitleSize = 0.5,
  facetBGColor = NA,
  facetLabelColor = "black",
  facetLabelSize = 1,
  facetLabelSizeMultiAB = 1,
  facetLabelBorderLwd = NA_real_,
  numeric2Cat_list = NULL,
  frameShow = F,
  pdfpng = "png",
  fillcolorNA = "gray",
  fillshowNA = NA,
  fillcolorNULL = "gray",
  legendSingleColorOn = T,
  legendSingleValue = NULL,
  legendSingleColor = "white",
  facetCols = 4,
  mapTitleOn = T,
  innerMargins = c(0, 0, 0, 0),
  legendDigitsOverride = NULL,
  classPalette = NULL,
  classPaletteDiff = "pal_div_BrGn"
)
}
\arguments{
\item{polygonDataTables}{Default = NULL,}

\item{gridDataTables}{Default = NULL,}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{folderName}{Default ="folderNameDefault",}

\item{xRange}{Default ="All",}

\item{labels}{Default = F,}

\item{labelsSize}{Default = 1.2,}

\item{subRegShape}{Default = NULL,}

\item{subRegShpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),}

\item{subRegShpFile}{Default = paste("gadm36_1",sep=""),}

\item{subRegCol}{Default ="subRegion",}

\item{dirNameAppend}{Default =""}

\item{nameAppend}{Default =""}

\item{legendOutsideSingle}{Default =F, Single plots by default have legends inside. This can be moved out if wanted.}

\item{legendOutsidePosition}{Default = NULL, # "right","left","top","bottom", "center"}

\item{legendPosition}{Default = NULL, # c("RIGHT','top') - RIGHT LEFT TOP BOTTOM}

\item{legendFixedBreaks}{Default = "5",}

\item{legendTitleSizeO}{Default = 2,}

\item{legendTextSizeO}{Default =1,}

\item{legendTitleSizeI}{Default = 1,}

\item{legendTextSizeI}{Default =0.5,}

\item{animateOn}{Default = T,}

\item{fps}{Default = 1,}

\item{scenRef}{Reference Scenario. Default = NULL}

\item{scenDiff}{Scenarios to Diff. Default = NULL}

\item{extension}{Default =F,}

\item{boundaryRegShape}{Default = NULL,}

\item{boundaryRegShpFolder}{Default= NULL . Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default="")}

\item{boundaryRegShpFile}{Default=NULL . Suggested paste("ne_10m_admin_0_countries",sep  Default=""),}

\item{boundaryRegCol}{Default=NULL. Suggested "NAME_0",}

\item{boundaryRegionsSelect}{Default = NULL,}

\item{extendedLabels}{Default = T}

\item{extendedFillColor}{Default ="grey75",}

\item{extendedBGColor}{Default ="lightblue1",}

\item{extendedHighLightColor}{Default ="cornsilk1",}

\item{extendedLabelsColor}{Default ="grey30",}

\item{extdendedLabelSize}{Default =0.7,}

\item{extendedShape}{Default =NULL,}

\item{extendedShapeCol}{Default =NULL,}

\item{expandPercent}{Default =2}

\item{projX}{Default = projX="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"}

\item{figWidth}{Default =9}

\item{figHeight}{Default =7}

\item{scaleRange}{Default NULL. A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{scaleRangeDiffAbs}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{scaleRangeDiffPrcnt}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{paramsSelect}{Default ="All"}

\item{multifacetsOn}{Default = F,}

\item{multiFacetCols}{Default ="multiFacetRow",}

\item{multiFacetRows}{Default ="multiFacetCol",}

\item{legendOutsideMulti}{Default = NULL,}

\item{legendPositionMulti}{Default = NULL,}

\item{legendTitleSizeMulti}{Default = NULL,}

\item{legendTextSizeAnim}{Default = NULL,}

\item{legendTextSizeMulti}{Default = NULL,}

\item{refMultiA}{Default = NULL , eg. "gfdl-esm2m"}

\item{refMultiB}{Default = NULL , eg. "rcp2p6"}

\item{chosenRefMeanYears}{Default=NULL}

\item{mapTitleSize}{Default=0.5}

\item{facetBGColor}{Default = NA}

\item{facetLabelColor}{Default = Black}

\item{facetLabelSize}{Default =2.5}

\item{facetLabelSizeMultiAB}{Default =1}

\item{facetLabelBorderLwd}{Default=NA_real_,}

\item{numeric2Cat_list}{Default=NULL,}

\item{frameShow}{Default = T. Whether to plot frame around maps and facets.}

\item{pdfpng}{Save IO figures as pdf or png. Type=String. Options: 'pdf' or 'png'. Default = 'png'}

\item{fillcolorNA}{Default="gray",}

\item{fillshowNA}{Default=NA,}

\item{fillcolorNULL}{Default="gray"}

\item{legendSingleColorOn}{Default=F,}

\item{legendSingleValue}{Default =NULL,}

\item{legendSingleColor}{Default="white"}

\item{facetCols}{Default=4}

\item{mapTitleOn}{Default=T}

\item{innerMargins}{Default =c(0,0,0,0) # bottom, left, top, right}

\item{legendDigitsOverride}{Default=NULL}

\item{classPalette}{Default = NULL}

\item{classPaletteDiff}{Default = "pal_div_BrGn"}
}
\value{
Returns the formatted data used to produce chart

A list with the gridTbl and shapeTbl used to plot the data if any.
}
\description{
This function produce different kinds of maps for the metis package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
