% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.prepGrid.R
\name{metis.prepGrid}
\alias{metis.prepGrid}
\title{metis.prepGrid}
\usage{
metis.prepGrid(demeterFolder = NULL, demeterScenario = NULL,
  demeterTimesteps = seq(from = 2005, to = 2100, by = 5),
  demeterUnits = NULL, tethysFolder = NULL, tethysScenario = NULL,
  tethysUnits = NULL, tethysFiles = c("wddom", "wdelec", "wdirr",
  "wdliv", "wdmfg", "wdmin", "wdnonag", "wdtotal"), xanthosFolder = NULL,
  xanthosScenario = NULL, xanthosUnits = NULL, xanthosFiles = NULL,
  xanthosCoordinatesPath = paste(getwd(),
  "/dataFiles/grids/xanthosCoords/coordinates.csv", sep = ""),
  scarcityXanthosRollMeanWindow = 10, dirOutputs = paste(getwd(),
  "/outputs", sep = ""), reReadData = 1,
  gridMetisData = paste(dirOutputs, "/Grids/gridMetis.RData", sep = ""))
}
\arguments{
\item{demeterFolder}{Full path to demeter outputs}

\item{demeterScenario}{Name of demeter scenario}

\item{demeterTimesteps}{Default is seq(from=2005,to=2100,by=5)}

\item{demeterUnits}{No Default}

\item{tethysFolder}{Folder for tethys results}

\item{tethysScenario}{Scenario name for tethys run}

\item{tethysUnits}{No Default}

\item{tethysFiles}{Default =c("wddom","wdelec","wdirr","wdliv","wdmfg","wdmin","wdnonag","wdtotal"),}

\item{xanthosFolder}{Xanthos Folder Path}

\item{xanthosScenario}{Xanthos Scenario Name}

\item{xanthosUnits}{Xanthos Untis}

\item{xanthosFiles}{Xanthos Files to Read}

\item{xanthosCoordinatesPath}{paste(getwd(),"/dataFiles/grids/xanthosCoords/coordinates.csv",sep="")}

\item{scarcityXanthosRollMeanWindow}{Default = 10,}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{reReadData}{Default =1,}

\item{gridMetisData}{Default = paste(dirOutputs, "/Grids/gridMetis.RData", sep = "")}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
