% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.printPdfPng.R
\name{metis.printPdfPng}
\alias{metis.printPdfPng}
\title{metis.printPdfPng}
\usage{
metis.printPdfPng(
  figure = NULL,
  dir = getwd(),
  filename = "plot",
  figWidth = 13,
  figHeight = 9,
  pdfpng = "png",
  transparent = F
)
}
\arguments{
\item{figure}{Default=NULL. Figure to be printed}

\item{dir}{Default = getwd(). Directory to print figure}

\item{filename}{Default = "plot". File name}

\item{figWidth}{Default=13.}

\item{figHeight}{Default=9.}

\item{pdfpng}{Default="png". Either "pdf" or "png"}

\item{transparent}{Default=F}
}
\value{
Prints out graphic
}
\description{
This function prints figure to pdf or png.
}
\keyword{charts,}
\keyword{diffplots}
