% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.runAll.R
\name{metis.runAll}
\alias{metis.runAll}
\title{metis.runAll}
\usage{
metis.runAll(run_readGCAM = F, run_charts = F, run_boundaries = F,
  run_bia = F, run_prepGrids = F, run_grid2poly = F, run_maps = F,
  run_io = F, gcamcountryNames = NULL, reReadDataGCAM = F,
  gcamdatabasePath = paste("D:/GCAM/gcam-core_LAC/output", sep = ""),
  gcamdatabaseName = "database_basexdb", dataProjPath = paste(getwd(),
  "/dataFiles/gcam", sep = ""), dataProj = paste("GCAMdataProj.proj", sep
  = ""), queryPath = paste(getwd(), "/dataFiles/gcam", sep = ""),
  queryxml = "metisQueries.xml",
  scenOrigNames = c("IDBUruguay_GCAMOrig", "IDBUruguay_GCAMRef"),
  scenNewNames = c("GCAMOrig", "GCAMRef"),
  paramsSelect_GCAM = c("All"), dataTablesLocal = NULL,
  scenRef = "GCAMRef", regionCompareOnly = 0,
  scenarioCompareOnly = 1, useNewLabels = 1,
  colOrder1 = c("GCAMOrig", "GCAMRef", "Local Data"),
  colOrderName1 = "scenario", xCompare = c("2015", "2030", "2050",
  "2100"), xRange = c(2010, 2015, 2020, 2025, 2030, 2035, 2040, 2045,
  2050), paramsSelect_Chart = c("All"), localcountryName = NULL,
  localShapeFileFolder = NULL, localShapeFile = NULL,
  localShapeFileColName = NULL, boundaryGridsOverlap = c(paste(getwd(),
  "/dataFiles/grids/emptyGrids/grid_025.csv", sep = ""), paste(getwd(),
  "/dataFiles/grids/emptyGrids/grid_050.csv", sep = "")),
  biaInputsFolder = paste(getwd(), "/dataFiles/grids/bia/biaInputs", sep
  = ""), biaInputsFiles = c("global_power_plant_database_MW"),
  gridChoice = "grid_025", diagnosticsON = T,
  subsectorNAdistribute = "totalOther", nameAppendBia = "",
  reReadDataPrepGrids = 0, demeterFolder = paste(getwd(),
  "/dataFiles/grids/demeter/", sep = ""), demeterScenario = "Eg1",
  demeterUnits = "Landuse (Fraction)", demeterTimesteps = seq(from =
  2005, to = 2020, by = 5), tethysFolder = paste(getwd(),
  "/dataFiles/grids/tethys/", sep = ""), tethysScenario = "Eg1",
  copySingleTethysScenbyXanthos = "Eg1", tethysFiles = c("wddom",
  "wdelec", "wdirr", "wdliv", "wdmfg", "wdmin", "wdnonag", "wdtotal"),
  tethysUnits = "Water Withdrawals (mm)",
  xanthosFolder = paste(getwd(), "/dataFiles/grids/xanthosRunsChris/",
  sep = ""),
  xanthosFiles = c("pm_abcd_mrtm_noresm1-m_rcp8p5_1950_2099/q_km3peryear_pm_abcd_mrtm_noresm1-m_rcp8p5_1950_2099.csv"),
  xanthosCoordinatesPath = paste(getwd(),
  "/dataFiles/grids/xanthosReference/coordinates.csv", sep = ""),
  xanthosGridAreaHecsPath = paste(getwd(),
  "/dataFiles/grids/xanthosReference/Grid_Areas_ID.csv", sep = ""),
  scarcityXanthosRollMeanWindow = 10, spanLowess = 0.25,
  popFolder = paste(getwd(), "/dataFiles/grids/griddedIDsPop/", sep =
  ""), popFiles = "grid_pop_map", popUnits = "Population",
  biaFolder = paste(getwd(), "/dataFiles/grids/bia/biaOutputs/", sep =
  ""), biaFiles = paste("dataBia", nameAppendBia, sep = ""),
  gridMetisData = paste(getwd(), "/outputs/Grids/gridMetis.RData", sep =
  ""), sqliteUSE = T, sqliteDBNamePath = paste(getwd(),
  "/outputs/Grids/gridMetis.sqlite", sep = ""), grid = NULL,
  paramsSelect_grid2poly = "All", polygonDataTablesCustom = NULL,
  sqliteUSEGrid2Poly = T, sqliteDBNamePathGrid2Poly = paste(getwd(),
  "/outputs/Grids/gridMetis.sqlite", sep = ""), run_state_grid2poly = T,
  run_GCAMbasin_grid2poly = T, run_localShape_grid2poly = T,
  scaleRange_i = NULL, xRangeMap = seq(from = 2000, to = 2050, by = 5),
  legendPosition = c("LEFT", "bottom"), legendOutsideSingle = T,
  animateOn = T, delay = 100, scenRefMap = "gfdl-esm2m_rcp2p6_NA_NA",
  paramsSelect_Map = c("All"), indvScenarios = "All",
  GCMRCPSSPPol = F, localSubRegDataTables = NULL, run_map_grid = T,
  run_map_state = T, run_map_GCAMbasin = T,
  run_map_localShapeLocalData = T, run_map_localShapeGCAMData = T,
  run_map_localShapeGCAMGrid = T)
}
\arguments{
\item{run_readGCAM}{Default = T,}

\item{run_charts}{Default = T,}

\item{run_boundaries}{Default = T,}

\item{run_bia}{Default = T,}

\item{run_prepGrids}{Default = T,}

\item{run_grid2poly}{Default = T,}

\item{run_maps}{Default = T,}

\item{run_io}{Default = T,}

\item{gcamcountryNames}{Default = c("Argentina","Colombia","Uruguay"),}

\item{reReadDataGCAM}{Default = F,}

\item{gcamdatabasePath}{Default =paste("D:/GCAM/gcam-core_LAC/output",sepDefault =""),}

\item{gcamdatabaseName}{Default ="database_basexdb",}

\item{dataProjPath}{Default = paste(getwd(),"/dataFiles/gcam",sepDefault =""),}

\item{dataProj}{Default = paste("GCAMdataProj.proj",sepDefault =""),}

\item{queryPath}{Default = paste(getwd(),"/dataFiles/gcam",sepDefault =""),}

\item{queryxml}{Default = "metisQueries.xml",}

\item{scenOrigNames}{Default = c("IDBUruguay_GCAMOrig", "IDBUruguay_GCAMRef"),}

\item{scenNewNames}{Default = c("GCAMOrig","GCAMRef"),}

\item{paramsSelect_GCAM}{Default = c("All"),}

\item{dataTablesLocal}{Default = NULL.}

\item{scenRef}{Default = "GCAMRef"}

\item{regionCompareOnly}{Default = 0, # Default is "0"}

\item{scenarioCompareOnly}{Default = 1, # Default is "0"}

\item{useNewLabels}{Default = 1,}

\item{colOrder1}{Default = c("GCAMOrig","GCAMRef","Local Data"),}

\item{colOrderName1}{Default = "scenario",}

\item{xCompare}{Default = c("2015","2030","2050","2100"),}

\item{xRange}{Default = c(2010,2015,2020,2025,2030,2035,2040,2045,2050),}

\item{paramsSelect_Chart}{Default = c("All"),}

\item{localcountryName}{Default = NULL}

\item{localShapeFileFolder}{Default = NULL}

\item{localShapeFile}{Default = "NULL}

\item{localShapeFileColName}{Default = NULL, # Make sure this is one of the names(tempShape@data)}

\item{boundaryGridsOverlap}{Default = NULL}

\item{biaInputsFolder}{Default = paste(getwd(),"/dataFiles/grids/bia/biaInputs",sepDefault =""),}

\item{biaInputsFiles}{Default = c("global_power_plant_database_MW"),}

\item{gridChoice}{Default = "grid_025", # Choose between "grid_025" and "grid_050"}

\item{diagnosticsON}{Default = T,}

\item{subsectorNAdistribute}{Default = "totalOther", # Choose between "totalOther" or "even"}

\item{nameAppendBia}{Default = "",}

\item{reReadDataPrepGrids}{Default =0,}

\item{demeterFolder}{Default =paste(getwd(),"/dataFiles/grids/demeter/",sepDefault =""),}

\item{demeterScenario}{Default ="Eg1",}

\item{demeterUnits}{Default ="Landuse (Fraction)",}

\item{demeterTimesteps}{Default =seq(fromDefault =2005,toDefault =2020,byDefault =5),}

\item{tethysFolder}{Default =paste(getwd(),"/dataFiles/grids/tethys/",sepDefault =""),}

\item{tethysScenario}{Default ="Eg1",}

\item{copySingleTethysScenbyXanthos}{Default ="Eg1",}

\item{tethysFiles}{Default =c("wddom","wdelec","wdirr","wdliv","wdmfg","wdmin","wdnonag","wdtotal"),}

\item{tethysUnits}{Default ="Water Withdrawals (mm)",}

\item{xanthosFolder}{Default =paste(getwd(),"/dataFiles/grids/xanthosRunsChris/",sepDefault =""),}

\item{xanthosFiles}{Default = See file in code. Name too long to print here.}

\item{xanthosCoordinatesPath}{Default =paste(getwd(),"/dataFiles/grids/xanthosReference/coordinates.csv",sepDefault =""),}

\item{xanthosGridAreaHecsPath}{Default =paste(getwd(),"/dataFiles/grids/xanthosReference/Grid_Areas D.csv",sepDefault =""),}

\item{scarcityXanthosRollMeanWindow}{Default =10,}

\item{spanLowess}{Default =0.25,}

\item{popFolder}{Default =paste(getwd(),"/dataFiles/grids/griddedIDsPop/",sepDefault =""),}

\item{popFiles}{Default ="grid_pop_map",}

\item{popUnits}{Default ="Population",}

\item{biaFolder}{Default =paste(getwd(),"/dataFiles/grids/bia/biaOutputs/",sepDefault =""),}

\item{biaFiles}{Default =paste("dataBia",nameAppendBia,sepDefault =""),}

\item{gridMetisData}{Default =paste(getwd(),"/outputs/Grids/gridMetis.RData", sep Default = ""),}

\item{sqliteUSE}{Default = T,}

\item{sqliteDBNamePath}{Default =paste(getwd(),"/outputs/Grids/gridMetis.sqlite", sep Default = ""),}

\item{grid}{Default = NULL,}

\item{paramsSelect_grid2poly}{Default = "All"}

\item{polygonDataTablesCustom}{Default = NULL,}

\item{sqliteUSEGrid2Poly}{Default = T,}

\item{sqliteDBNamePathGrid2Poly}{Default = paste(getwd(),"/outputs/Grids/gridMetis.sqlite", sep Default = ""),}

\item{run_state_grid2poly}{Default = F,}

\item{run_GCAMbasin_grid2poly}{Default = F,}

\item{run_localShape_grid2poly}{Default = T,}

\item{scaleRange_i}{Default = NULL,}

\item{xRangeMap}{Default = seq(fromDefault =2000,toDefault =2050,byDefault =5),}

\item{legendPosition}{Default =c("LEFT","bottom"),}

\item{legendOutsideSingle}{Default =T,}

\item{animateOn}{Default =T,}

\item{delay}{Default =100,}

\item{scenRefMap}{Default ="gfdl-esm2m_rcp2p6_NA_NA",}

\item{paramsSelect_Map}{Default = c("All"),}

\item{indvScenarios}{Default = "All",}

\item{GCMRCPSSPPol}{Default =F,}

\item{localSubRegDataTables}{Default = NULL}

\item{run_map_grid}{Default = T,}

\item{run_map_state}{Default = T,}

\item{run_map_GCAMbasin}{Default = T,}

\item{run_map_localShapeLocalData}{Default = T}

\item{run_map_localShapeGCAMData}{Default = T}

\item{run_map_localShapeGCAMGrid}{Default = T}
}
\value{
A tibble with GCAM electricity generation distributed on a grid for selected regions
}
\description{
This function is the wrapper to run the full metis workflow. Defaults can be changed to modifiy the runs.
modules to Run
}
\keyword{metis}
