% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictPs}
\alias{predictPs}
\title{predictPs}
\usage{
predictPs(psModel, population, cohortMethodData)
}
\arguments{
\item{psModel}{An object of propensity score model}

\item{population}{The population data of target}

\item{cohortMethodData}{The cohort method data of target}
}
\value{
A dataframe containing the prediction for each person in the populatio.
}
\description{
Predict the propensity score of given population using external propensity score model
}
\details{
The function applied the trained model on the plpData to make predictions
}
