% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.assumptions.R
\name{srn.assumptions}
\alias{srn.assumptions}
\title{srn.assumptions}
\usage{
srn.assumptions()
}
\value{
A list of assumptions
}
\description{
This function loads holds the different assumptions used throughout the srn package.
}
\details{
List of Assumptions
\itemize{
\item convEJ2TWh
\item convEJ2GW
\item conv1975USDperGJ22017USDperMWh
\item conv1975USDperGJ22017USDperMBTU
\item convertGgTgMTC
\item GWPType}
}
\examples{
library(srn)
a<-srn.assumptions()
a # will give full list of assumptions
}
\keyword{assumptions}
