% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.colors.R
\name{srn.colors}
\alias{srn.colors}
\title{srn.colors}
\usage{
srn.colors(palx = NULL)
}
\arguments{
\item{palx}{Palette name to view the palette colors. Eg. srn.colors("pal_Basic")}
}
\value{
A list of color palettes.
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for SRN modeling to the global environment
}
\details{
List of Color Palettes
\itemize{
\item pal_HDDCDD
\item pal_16
\item elec_tech_colors
\item elec_renew_colors
\item building_colors
\item trn_fuel_colors
\item enduse_fuel_numbered
\item enduse_colors
\item pal_pri_ene
\item pal_pri_fuelcost
\item pal_emiss_sector
\item pal_landuse
\item pal_hydrogen
\item pal_refliq
\item emiss_by_enduse_colors
\item biouse_colors
\item pal_Basic
\item pal_Gas
\item pal_Diff
\item pal_Diff5
\item pal_Absolute
\item pal_Absolute5
\item pal_Unassigned
\item pal_elec_subsec
\item pal_elec_finalNrgFuel
\item pal_elec_techs
\item pal_elec_sec
\item pal_finalNrg_sec
\item pal_pri_ene
\item pal_elec_tech_colors}
}
\examples{
library(srn)
a<-srn.colors()
pie(rep(1,length(a$pal_Basic)),label=names(a$pal_Basic),col=a$pal_Basic)
}
\keyword{colors,}
\keyword{palette}
