% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.mapsProcess.R
\name{srn.mapProcess}
\alias{srn.mapProcess}
\title{srn.mapProcess}
\usage{
srn.mapProcess(polygonDataTables = NULL, gridDataTables = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), xRange = "All",
  labels = F, labelsSize = 1.2, regionsSelect = NULL,
  subRegShape = NULL, subRegShpFolder = paste(getwd(),
  "/dataFiles/gis/admin_gadm36", sep = ""),
  subRegShpFile = paste("gadm36_1", sep = ""), subRegCol = "NAME_1",
  subRegType = "subRegType", aggType = NULL, nameAppend = "",
  rasterCoverNegShape = T, legendOutsidePosition = NULL,
  legendPosition = NULL)
}
\arguments{
\item{polygonDataTables}{Default = NULL,}

\item{gridDataTables}{Default = NULL,}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{xRange}{Default ="All",}

\item{labels}{Default = F,}

\item{labelsSize}{Default = 1.2,}

\item{regionsSelect}{Default = NULL,}

\item{subRegShape}{Default = NULL,}

\item{subRegShpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),}

\item{subRegShpFile}{Default = paste("gadm36_1",sep=""),}

\item{subRegCol}{Default ="NAME_1",}

\item{subRegType}{Default ="subRegType",}

\item{aggType}{Default = NULL,}

\item{nameAppend}{Default =""}

\item{rasterCoverNegShape}{Default =T}

\item{legendOutsidePosition}{Default = NULL, # "right","left","top","bottom", "center"}

\item{legendPosition}{Default = NULL, # c("RIGHT','top') - RIGHT LEFT TOP BOTTOM}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the srn package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
