% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.prepGrid.R
\name{srn.prepGrid}
\alias{srn.prepGrid}
\title{srn.prepGrid}
\usage{
srn.prepGrid(demeterFolder, demeterScenario, demeterTimesteps = seq(from
  = 2005, to = 2100, by = 5), dirOutputs = paste(getwd(), "/outputs", sep
  = ""), reReadData = 1, gridSRNData = paste(dirOutputs,
  "/Grids/gridSRN.RData", sep = ""))
}
\arguments{
\item{demeterFolder}{Full path to demeter outputs}

\item{demeterScenario}{Name of demeter scenario}

\item{demeterTimesteps}{Default is seq(from=2005,to=2100,by=5)}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{reReadData}{Default =1,}

\item{gridSRNData}{Default = paste(dirOutputs, "/Grids/gridSRN.RData", sep = "")}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other srn modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
