% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.readgcam.R
\name{srn.readgcam}
\alias{srn.readgcam}
\title{srn.readgcam}
\usage{
srn.readgcam(gcamdatabase, queryxml)
}
\arguments{
\item{gcamdatabase}{Full path to gcam database}

\item{queryxml}{Full path to query.xml file}
}
\value{
A table with data formatted for srn package charting functions
}
\description{
This function connects to a gcamdatabase and uses a query file to
out results into a table ready for plotting.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
