% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.templates.R
\name{srn.templates}
\alias{srn.templates}
\alias{srn.printPdfPng}
\alias{srn.chartsThemeLight}
\alias{srn.tmapAnimate}
\alias{srn.tmapLayout}
\title{srn.templates}
\usage{
srn.printPdfPng(figure, dir, filename, figWidth, figHeight,
  pdfpng = "png")

srn.chartsThemeLight()

srn.tmapAnimate(map, filename = "animation.gif", width, height,
  delay = 60)

srn.tmapLayout()
}
\arguments{
\item{figure}{Figure to be printed in function srn.printPdfPng}

\item{dir}{Directory to print figure to in function srn.printPdfPng}

\item{filename}{Filename for figure printed in function srn.printPdfPng}

\item{figWidth}{Figure Width in inches for figures to be printed in function srn.printPdfPng}

\item{figHeight}{Figure height in inches for figures to be printed in function srn.printPdfPng}

\item{pdfpng}{Either "pdf", "png" or "both" to define the format of output}
}
\value{
A list of different templates
}
\description{
This script holds various templates used for different scripts.
}
\details{
List of Templates in this script:
\itemize{
\item srn.printPdfPng: Function used to print charts to a pdf or png or both.
\item srn.chartsThemeLight: A light ggplot theme for charts
\item srn.tmapAnimate: A function to animate tmaps across a variable.
\item srn.tmapLayout: A fucntion to define tmap layouts
}
}
\keyword{charts,}
\keyword{maps,}
\keyword{print}
\keyword{templates,}
