% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis-package.R
\name{use_credentials}
\alias{use_credentials}
\alias{read_credentials}
\title{Use Credentials from .aws/credentials File}
\description{
Use Credentials from .aws/credentials File
}
\references{
\code{\link[aws.signature:use_credentials]{aws.signature::use_credentials()}} /  \code{\link[aws.signature:read_credentials]{aws.signature::read_credentials()}}
}
