% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xmlMetisQueries}
\alias{xmlMetisQueries}
\title{metisQueries xml file}
\format{
.xml
}
\source{
Metis
}
\usage{
xmlMetisQueries
}
\description{
metisQueries xml file
}
\examples{
\dontrun{
 library(metis); library(XML)
 metis::xmlMetisQueries
 # Can save xml
 XML::saveXML(metis::xmlMetisQueries, file=paste(getwd(), "/metisQueries.xml", sep = ""))
}
}
\keyword{datasets}
