% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbConnect,AthenaDriver-method}
\alias{dbConnect,AthenaDriver-method}
\title{AthenaJDBC}
\usage{
\S4method{dbConnect}{AthenaDriver}(drv, Schema = "default",
  AwsRegion = "us-east-1",
  AwsCredentialsProviderClass = "com.simba.athena.amazonaws.auth.DefaultAWSCredentialsProviderChain",
  S3OutputLocation = Sys.getenv("AWS_S3_STAGING_DIR", unset = ""),
  MaxErrorRetry = 10, ConnectTimeout = 10000, SocketTimeout = 10000,
  LogPath = "", LogLevel = 0, fetch_size = 1000L, ...)
}
\arguments{
\item{drv}{driver}

\item{Schema}{The name of the database schema to use when a schema is not explicitly
specified in a query. You can still issue queries on other schemas by explicitly
specifying the schema in the query.}

\item{AwsRegion}{AWS region the Athena tables are in}

\item{AwsCredentialsProviderClass}{JDBC auth provider; You can add a
lengrh1 character vecrtor named parameter `AwsCredentialsProviderArguments`
to the `dbConnect()`  call to use alternate auth providers. Use a
comma-separated list of String arguments.}

\item{S3OutputLocation}{A write-able bucket on S3 that you have permissions for}

\item{MaxErrorRetry, ConnectTimeout, SocketTimeout}{technical connection info that you should only muck with if you know what you're doing.}

\item{LogPath, LogPath}{The Athena JDBC driver can (shockingly) provide a decent bit
of data in logs. Set this to a temporary directory or something log4j can use. For
`LogPath` use the names ("INFO", "DEBUG", "WARN", "ERROR", "ALL", "OFF", "FATAL", "TRACE") or
their corresponding integer values 0-6.}

\item{fetch_size}{Athena results fetch size}

\item{...}{passed on to the driver. See Details.}
}
\description{
Connect to Athena
}
\details{
Mandatory JDBC connection parameters are also named function
parameters. You can use `...` to supply additional/optional
parameters.
}
\section{Higlighted Extra Driver Configuration Options}{


These are take from the second item in References. See that resource
for more information.

- `BinaryColumnLength`: <int> The maximum data length for `BINARY` columns. Default `32767L`
- `ComplexTypeColumnLength`: <int> The maximum data length for `ARRAY`, `MAP`, and `STRUCT` columns. Default `65535L`
- `StringColumnLength`: <int> The maximum data length for `STRING` columns. Default `255L`
}

\references{
[Connect with JDBC](https://docs.aws.amazon.com/athena/latest/ug/connect-with-jdbc.html);
    [Simba Athena JDBC Driver with SQL Connector Installation and Configuration Guide](https://s3.amazonaws.com/athena-downloads/drivers/JDBC/SimbaAthenaJDBC_2.0.6/docs/Simba+Athena+JDBC+Driver+Install+and+Configuration+Guide.pdf)
}
