% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS49}
\alias{mapGCAMBasinsUS49}
\title{GCAM Basins cropped to US49}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from metis maps.
}
\usage{
mapGCAMBasinsUS49
}
\description{
GCAM Basins cropped to US49
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMBasinsUS49)
 head(mapGCAMBasinsUS49@data)
}
}
\keyword{datasets}
