% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS52}
\alias{mapGCAMBasinsUS52}
\title{GCAM Basins cropped to US52}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from metis maps
}
\usage{
mapGCAMBasinsUS52
}
\description{
GCAM Basins cropped to US52
}
\examples{
\dontrun{
 library(sp); library(metis)
 sp::plot(mapGCAMBasinsUS52)
 head(mapGCAMBasinsUS52@data)
}
}
\keyword{datasets}
