% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS52}
\alias{mapGCAMBasinsUS52}
\title{Cropped of GCAM Basins and US52}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMBasinsUS52
}
\description{
Cropped of GCAM Basins and US52
}
\examples{
\dontrun{
library(metis)
head(mapGCAMBasinsUS52)
}
}
\keyword{datasets}
