% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.bia.R
\name{metis.bia}
\alias{metis.bia}
\title{metis.bia}
\usage{
metis.bia(biaInputsFolder = "NA", biaInputsFiles = "NA",
  biaScenarioAssign = "NA", reReadData = 1, regionsSelect = NULL,
  dataProj = "dataProj.proj", dataProjPath = gcamdatabasePath,
  scenOrigNames = c("GCAMOrig", "GCAMModified"),
  scenNewNames = c("GCAMOrig", "GCAMModified"),
  gcamdatabasePath = "NA", gcamdatabaseName = "NA",
  queryxml = "metisQueries.xml", queryPath = gcamdatabasePath,
  queriesSelect = "All", paramsSelect = c("elecByTech",
  "elecCapBySubsector"), gridChoice = "grid_050", diagnosticsON = T,
  subsectorNAdistribute = "even")
}
\arguments{
\item{biaInputsFolder}{andym Bia Inputs Folder Path}

\item{biaInputsFiles}{andym Bia Files to Read}

\item{biaScenarioAssign}{andym Default "NA". Scenario name if testing a single scenario.}

\item{reReadData}{Default =1,}

\item{regionsSelect}{The regions to analyze in a vector. Example c('Colombia','Argentina')}

\item{dataProj}{Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.}

\item{dataProjPath}{Folder that contains the dataProj or where it will be produced.}

\item{scenOrigNames}{Original Scenarios names in GCAM database in a string vector.
For example c('scenario1','scenario2).}

\item{scenNewNames}{New Names which may be shorter and more useful for figures etc.
Default will use Original Names. For example c('scenario1','scenario2)}

\item{gcamdatabasePath}{Path to gcam database folder}

\item{gcamdatabaseName}{Name of gcam database}

\item{queryxml}{Full path to query.xml file}

\item{queryPath}{Folder that contains the query.xml file.By default it is
the same folder as specified by gcamdatabasePath}

\item{queriesSelect}{Default = "All". Vector of queries to read from the queryxml for example}

\item{paramsSelect}{Default = c("elecByTech", "elecCapBySubsector") . Vector of parameters to be read from the GCAM database}

\item{gridChoice}{Default = "grid_050" . Choice of whether to use 50 km x 50 km grid cells ("grid_050") or 25 km x 25 km ("grid_025").}

\item{diagnosticsON}{Default = T.}

\item{subsectorNAdistribute}{Default = "even". Choose "even" for even distribution or "totalOther" to distribute based on sum of all other subsectors..}

\item{biaOutputsFolder}{Default =paste(getwd(),"/dataFiles/grids/bia/biaOutputs",sep=""),}

\item{reReadData}{If TRUE will read the GCAM data base and create a queryData.proj file
in the same folder as the GCAM database. If FALSE will load a '.proj' file if a file
with full path is provided otherwise it will search for a dataProj.proj file in the existing
folder which may have been created from an old run.}
}
\value{
#andym a tibble with GCAM electricity generation distributed on a grid for a selected region
}
\description{
This function downscales GCAM electricity generation and installed capacity onto a grid, based on WRI PowerWatch dataset of present capacity
}
\keyword{downscale,}
\keyword{downscaled}
\keyword{downscaling,}
\keyword{electricity,}
\keyword{gcam,}
\keyword{generation,}
\keyword{gridded,}
