% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.chart.R
\name{metis.chart}
\alias{metis.chart}
\title{metis.chart}
\usage{
metis.chart(data, chartType = "bar", position = "stack", xData = "x",
  yData = "value", class = "class1", group = "scenario",
  classPalette = "classPalette1", classLabel = "classLabel1",
  xLabel = "xLabel", yLabel = "yLabel", facet_rows = "region",
  facet_columns = "scenario", ncolrow = 4, scales = "fixed",
  useNewLabels = 0, units = "units", xBreaksMaj = 10,
  xBreaksMin = 5, yBreaksMajn = 5, yBreaksMinn = 10,
  sizeBarLines = 0.5, sizeLines = 1.5, printFig = T,
  fileName = "chart", dirOutputs = paste(getwd(), "/outputs", sep =
  ""), figWidth = 13, figHeight = 9, pdfpng = "png")
}
\arguments{
\item{data}{data table for charting}

\item{chartType}{Type of chart: "bar" or "line"}

\item{position}{Position in bar charts. "identity", "stack" or "dodge"}

\item{xData}{Default "x"}

\item{yData}{Default "value"}

\item{class}{Default "class1"}

\item{group}{Default "scenario"}

\item{classPalette}{Default "classPalette1"}

\item{classLabel}{Default "classLabel1"}

\item{xLabel}{Default "xLabel"}

\item{yLabel}{Default "units"}

\item{facet_rows}{Default "region"}

\item{facet_columns}{Default "scenario"}

\item{ncolrow}{Number of columns or Rows for Faceted plots}

\item{scales}{Default "fixed"}

\item{useNewLabels}{Default 0}

\item{units}{Default "units"}

\item{xBreaksMaj}{Default 10}

\item{xBreaksMin}{Default 5}

\item{yBreaksMajn}{Default 5}

\item{yBreaksMinn}{Default 10}

\item{sizeBarLines}{Default 0.5}

\item{sizeLines}{Default 1.5}

\item{printFig}{Default = T,}

\item{fileName}{Default = "map",}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep Default = "")}

\item{figWidth}{Default = 9,}

\item{figHeight}{Default = 7,}

\item{pdfpng}{Default = "png",}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of charts for the metis package.
iIt requires a table in the Metis format. Each figure is accompanied with a csv table.
}
\examples{
# Examples below show the default chart with minimum information
# and then adding progressively more details.

library(tibble)
library(dplyr)
tbl <- tribble (
~x,     ~value,
2010,   15,
2020,   20,
2030,   30
)
metis.chart(data=tbl,xData="x",yData="value",chartType = "line")
metis.chart(data=tbl,xData="x",yData="value",chartType = "bar")
}
\keyword{charts,}
\keyword{diffplots}
