% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.colors.R
\name{metis.colors}
\alias{metis.colors}
\title{metis.colors}
\usage{
metis.colors(palx = NULL)
}
\arguments{
\item{palx}{Palette name to view the palette colors. Eg. metis.colors("pal_Basic")}
}
\value{
A list of color palettes.
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for Metis modeling to the global environment
}
\details{
List of Color Palettes
\itemize{
\item "pal_metis"
\item "pal_16"
\item "pal_Basic"
\item "pal_hot"
\item "pal_wet"
\item "pal_div_wet"
\item "pal_div_BlRd"
\item "pal_div_RdBl"
\item "pal_div_BrGn"
\item "pal_div_GnBr"
\item "pal_div_BluRd"
\item "pal_div_RdBlu"
\item "pal_green"
\item "pal_div_BrGn"
\item "pal_div_BlRd"
\item "pal_sankey"
\item "pal_spectral"
\item "pal_ScarcityCat"}
}
\examples{
library(metis)
a<-metis.colors()
pie(rep(1,length(a$pal_Basic)),label=names(a$pal_Basic),col=a$pal_Basic)
}
\keyword{colors,}
\keyword{palette}
