% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.grid2poly.R
\name{metis.grid2poly}
\alias{metis.grid2poly}
\title{metis.grid2poly}
\usage{
metis.grid2poly(grid = NULL, subRegShape = NULL,
  subRegShpFolder = NULL, subRegShpFile = NULL, subRegCol = NULL,
  subRegType = "subRegType", aggType = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "",
  labelsSize = 1.2, paramsSelect = "All", sqliteUSE = F,
  sqliteDBNamePath = paste(getwd(), "/outputs/Grids/gridMetis.sqlite",
  sep = ""))
}
\arguments{
\item{grid}{Default=NULL. Grid file in .csv format or a R table, data frame or tibble with as a minimum columns with "lat","lon" and "value",}

\item{subRegShape}{Default=NULL. shapefile over which grid data is to be aggregated.}

\item{subRegShpFolder}{Default=NULL. Folder containing boundary region shapefile. Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default=""),}

\item{subRegShpFile}{Default=NULL. Name of sub-region shapefile. Suggested paste("ne_10m_admin_1_states_provinces",sep  Default=""),}

\item{subRegCol}{Default= NULL. Suggested for states "name",}

\item{subRegType}{Default="subRegType". Eg. "states", "basins" etc.}

\item{aggType}{Default=NULL. Aggregation method to be used. Either "vol" or "depth" dependening on the type of data provided.}

\item{dirOutputs}{Default=paste(getwd(),"/outputs",sep  Default=""),}

\item{nameAppend}{Default="",}

\item{labelsSize}{Default =1.2. Label size for the region names for the gridoverlay plot.}

\item{paramsSelect}{Default ="All"}

\item{sqliteUSE}{Default = T,}

\item{sqliteDBNamePath}{Default = paste(getwd(),"/outputs/Grids/gridMetis.sqlite", sep = "")}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function takes a .csv file with gridded lat, long data and aggregates
the data by spatial boundaries given different shapefiles.
}
\examples{
   # Example Grid File (csv with lats and lons that overlap the shapefile)
       gridExample<-paste(getwd(),"/dataFiles/examples/example_grid_ArgentinaBermejo3_Eg1Eg2.csv",sep="")

       # Polygons. An example Shapefile is Provided with metis in ./metis/dataFiles/examples.
       examplePolyFolder_i<-paste(getwd(),"/dataFiles/examples",sep="")
       examplePolyFile_i<-paste("bermejo3Cropped",sep="")
       bermejo3Cropped=readOGR(dsn=examplePolyFolder,
                               layer=examplePolyFile,use_iconv=T,encoding='UTF-8')
       head(bermejo3Cropped@data) # TO choose subRegCol name
       subRegCol_i = "SUB_NAME"

       # Run metis.grid2poly
       polyBermeo3Cropped<-metis.grid2poly(grid=gridExample,
                                           subRegShpFolder=examplePolyFolder_i,
                                           subRegShpFile=examplePolyFile_i,
                                           subRegCol=subRegCol_i,
                                           aggType="depth", # Aggregation type. See ?metis.grid2poly
                                           nameAppend="_Bermeo3")
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
