% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.grid2poly.R
\name{metis.grid2poly}
\alias{metis.grid2poly}
\title{metis.grid2poly}
\usage{
metis.grid2poly(
  gridFiles = NULL,
  regionName = "region",
  subRegShape = NULL,
  subRegShpFolder = NULL,
  subRegShpFile = NULL,
  subRegCol = NULL,
  subRegType = "subRegType",
  aggType = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  folderName = NULL,
  nameAppend = "",
  labelsSize = 1.2,
  paramsSelect = "All",
  scenariosSelect = "All",
  paramScenariosFixed = NULL,
  tethysFilesScarcity = NULL,
  xanthosFilesScarcity = NULL,
  calculatePolyScarcity = F,
  calculatePolyScarcityOnly = F,
  saveFiles = T
)
}
\arguments{
\item{gridFiles}{Default=NULL. Grid file in .csv format or a R table, data frame or tibble with as a minimum columns with "lat","lon" and "value",}

\item{regionName}{Default = "region"}

\item{subRegShape}{Default=NULL. shapefile over which grid data is to be aggregated.}

\item{subRegShpFolder}{Default=NULL. Folder containing boundary region shapefile. Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default=""),}

\item{subRegShpFile}{Default=NULL. Name of sub-region shapefile. Suggested paste("ne_10m_admin_1_states_provinces",sep  Default=""),}

\item{subRegCol}{Default= NULL. Suggested for states "name",}

\item{subRegType}{Default="subRegType". Eg. "states", "basins" etc.}

\item{aggType}{Default=NULL. Aggregation method to be used. Either "vol" or "depth" dependening on the type of data provided.}

\item{dirOutputs}{Default=paste(getwd(),"/outputs",sep  Default=""),}

\item{folderName}{Default = NULL}

\item{nameAppend}{Default="",}

\item{labelsSize}{Default =1.2. Label size for the region names for the gridoverlay plot.}

\item{paramsSelect}{Default ="All"}

\item{scenariosSelect}{Default ="All"}

\item{paramScenariosFixed}{Default=NULL}

\item{tethysFilesScarcity}{Default =NULL,}

\item{xanthosFilesScarcity}{Default =NULL}

\item{calculatePolyScarcity}{Default = F}

\item{calculatePolyScarcityOnly}{Default = F}

\item{saveFiles}{Default =T}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function takes a .csv file with gridded lat, long data and aggregates
the data by spatial boundaries given different shapefiles.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
