% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.gridByPoly.R
\name{metis.gridByPoly}
\alias{metis.gridByPoly}
\title{metis.gridByPoly}
\usage{
metis.gridByPoly(grid = NULL, boundaryRegShpFolder = NULL,
  boundaryRegShpFile = NULL, colName = NULL,
  outputDir = paste(getwd(), "/outputs", sep = ""),
  fname = "gridByPoly", saveFile = F)
}
\arguments{
\item{grid}{Default = NULL. Full path to grid file.}

\item{boundaryRegShpFolder}{Default = NULL,}

\item{boundaryRegShpFile}{Default = NULL,}

\item{colName}{Default = NULL,}

\item{outputDir}{Default = paste(getwd(),"/outputs",sep=""),}

\item{fname}{Default = "gridByPoly"}

\item{saveFile}{Default = F. If want csv output then change to T}
}
\value{
Prints out graphic
}
\description{
This function finds the grids located within a given shapefiles regions
}
\keyword{grid,}
\keyword{polygon}
\keyword{shape,}
