% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.hydroInvest.R
\name{metis.hydroInvest}
\alias{metis.hydroInvest}
\title{metis.hydroInvest}
\usage{
metis.hydroInvest(addition_costs, start_year, end_year)
}
\arguments{
\item{addition_costs}{list formatted as produced by metis.elecInvest.R function}
}
\value{
Returns data in a form required by metislreadgcam.R
}
\description{
Function that calculates electricity subsector investment requirements from a given GCAM run.
}
\keyword{infrastructure}
\keyword{investments,}
