% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.io.R
\name{metis.io}
\alias{metis.io}
\title{metis.io}
\usage{
metis.io(Z0 = NULL, D0 = NULL, X0 = NULL, A0 = NULL,
  priorityZvsA = c(Z0 = 1, A0 = 2), priorityXvsCap = c(X0 = 1, Cap0 =
  2), Import0 = NULL, Export0 = NULL, Cap0 = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "")
}
\arguments{
\item{Z0}{Initial Nexus Flows (i.e. Supply sectors which also have demands).Default = NULL,}

\item{D0}{Intiial Other flows. (All other sectors which have demands but do not supply resources). Default = NULL,}

\item{X0}{Initial Total Demands. Default = NULL,}

\item{A0}{Initial Intensity Matrix. Default = NULL,}

\item{priorityZvsA}{Default = c(Z0=1, A0=2),}

\item{priorityXvsCap}{Default = c(X0=1, Cap0=2)}

\item{Import0}{Default =NULL,}

\item{Export0}{Default =NULL,}

\item{Cap0}{Capacity. Default =NULL,}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{nameAppend}{Modified intensity matrix. Default =NULL,}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with domestic metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
