% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.io.R
\name{metis.io}
\alias{metis.io}
\title{metis.io}
\usage{
metis.io(dirOutputs = paste(getwd(), "/outputs", sep = ""))
}
\arguments{
\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
