% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.io.R
\name{metis.io}
\alias{metis.io}
\title{metis.io}
\usage{
metis.io(ioTable0 = NULL, useIntensity = 0, A0 = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "")
}
\arguments{
\item{ioTable0}{Initial ioTable. Must have columns: supplySubSector,total,export and cap. Each supply sector should also have imports. Default = NULL,}

\item{useIntensity}{Boolean to use given intensity or not. Default is set to 0.}

\item{A0}{Intensity matrix. Default Null.}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{nameAppend}{Modified intensity matrix. Default =NULL,}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with domestic metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
