% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.io.R
\name{metis.io}
\alias{metis.io}
\title{metis.io}
\usage{
metis.io(ioTable0 = NULL, useIntensity = 0, A0 = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), nameAppend = "",
  figWidth = 9, figHeight = 7, sankeyLabelAbsPlots = 1,
  combSubRegionPlots = 1, folderName = "folderNameDefault",
  pdfpng = "png", sankeyAxis1Label = "From", sankeyAxis2Label = "To",
  plotSankeys = T)
}
\arguments{
\item{ioTable0}{Initial ioTable. Must have columns: supplySubSector,total,export and cap. Each supply sector should also have imports. Default = NULL,}

\item{useIntensity}{Boolean to use given intensity or not. Default is set to 0.}

\item{A0}{Intensity matrix. Default Null.}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{nameAppend}{Modified intensity matrix. Default =NULL,}

\item{figWidth}{Default = 9,}

\item{figHeight}{Default = 7,}

\item{sankeyLabelAbsPlots}{Default = 1}

\item{combSubRegionPlots}{Default = 1}

\item{folderName}{Default ="folderNameDefault"}

\item{pdfpng}{Save IO figures as pdf or png. Type=String. Options: 'pdf' or 'png'. Default = 'png'.}

\item{sankeyAxis1Label}{Set left-hand side label of sankey diagram. Type=String. Default = 'From'.}

\item{sankeyAxis2Label}{Set right-hand side label of sankey diagram. Type=String. Default = 'To'.}

\item{plotSankeys}{Default = T}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with domestic metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
