% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis-lite-package.R
\docType{package}
\encoding{UTF-8}
\name{metis.lite}
\alias{metis.lite}
\alias{metis.lite-package}
\title{Helpers for Accessing and Querying Amazon Athena}
\description{
Methods are provides to connect to 'Amazon' 'Athena', lookup schemas/tables,
perform queries and retrieve query results. A lightweight 'RJDBC' implementation
is included along with an interface to the 'AWS' command-line utility.
}
\section{IMPORTANT}{


Since R 3.5 (I don't remember this happening in R 3.4.x) signals sent from interrupting
Athena JDBC calls crash the R #' interpreter. You need to set the `-Xrs` option to avoid
signals being passed on to the JVM owner. That has to be done _#' before_ `rJava` is
loaded so you either need to remember to put it at the top of all scripts _or_ stick this
in your local #' `~/.Rprofile` and/or sitewide `Rprofile`:



Since R 3.5 (I don't remember this happening in R 3.4.x) signals sent from interrupting
Athena JDBC calls crash the R #' interpreter. You need to set the `-Xrs` option to avoid
signals being passed on to the JVM owner. That has to be done _#' before_ `rJava` is
loaded so you either need to remember to put it at the top of all scripts _or_ stick this
in your local #' `~/.Rprofile` and/or sitewide `Rprofile`:


    if (!grepl("-Xrs", getOption("java.parameters", ""))) {
      options(
        "java.parameters" = c(getOption("java.parameters", default = NULL), "-Xrs")
      )
    }
}

\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
