% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mappings.R
\name{metis.mappings}
\alias{metis.mappings}
\title{metis.mappings}
\usage{
metis.mappings(name = NULL)
}
\arguments{
\item{name}{Default=NULL. Name of assumption object.
\itemize{
\item mapParamQuery}}
}
\value{
A list of assumptions
}
\description{
This function has all the mappings used to convert between different categories
}
\details{
List of Mappings
}
\examples{
library(metis)
a<-metis.mappings()
a # will give full list of mappings
}
\keyword{assumptions}
