% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mapsProcess.R
\name{metis.mapProcess}
\alias{metis.mapProcess}
\title{metis.mapProcess}
\usage{
metis.mapProcess(polygonDataTables = NULL, gridDataTables = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), xRange = "All",
  labels = F, labelsSize = 1.2, subRegShape = NULL,
  subRegShpFolder = NULL, subRegShpFile = NULL, subRegCol = NULL,
  subRegType = "subRegType", nameAppend = "",
  legendOutsideSingle = F, legendOutsidePosition = NULL,
  legendPosition = NULL, legendFixedBreaks = 5, legendTitleSizeO = 2,
  legendTextSizeO = 1, legendTitleSizeI = 1.5, legendTextSizeI = 1,
  animateOn = T, delay = 100, scenRef = NULL, extension = F,
  boundaryRegShape = NULL, boundaryRegShpFolder = NULL,
  boundaryRegShpFile = NULL, boundaryRegCol = NULL,
  boundaryRegionsSelect = NULL, fillcolorNA = NULL,
  extendedFillColor = "grey75", extendedBGColor = "lightblue1",
  extendedHighLightColor = "cornsilk1", extendedLabelsColor = "grey30",
  extdendedLabelSize = 0.7, extendedShape = NULL,
  extendedShapeCol = NULL, expandPercent = 2,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  figWidth = 9, figHeight = 7, scaleRange = NULL,
  paramsSelect = "All", indvScenarios = T, GCMRCPSSPPol = F,
  multiFacetCols = "scenarioRCP", multiFacetRows = "scenarioGCM",
  legendOutsideMulti = T, legendPositionMulti = NULL,
  legendTitleSizeMulti = NULL, legendTextSizeAnim = NULL,
  legendTextSizeMulti = NULL, refGCM = NULL, refRCP = NULL,
  chosenRefMeanYears = NULL, mapTitleSize = 0.5)
}
\arguments{
\item{polygonDataTables}{Default = NULL,}

\item{gridDataTables}{Default = NULL,}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{xRange}{Default ="All",}

\item{labels}{Default = F,}

\item{labelsSize}{Default = 1.2,}

\item{subRegShape}{Default = NULL,}

\item{subRegShpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),}

\item{subRegShpFile}{Default = paste("gadm36_1",sep=""),}

\item{subRegCol}{Default ="NAME_1",}

\item{subRegType}{Default ="subRegType",}

\item{nameAppend}{Default =""}

\item{legendOutsideSingle}{Default =F, Single plots by default have legends inside. This can be moved out if wanted.}

\item{legendOutsidePosition}{Default = NULL, # "right","left","top","bottom", "center"}

\item{legendPosition}{Default = NULL, # c("RIGHT','top') - RIGHT LEFT TOP BOTTOM}

\item{legendFixedBreaks}{Default = "5",}

\item{legendTitleSizeO}{Default = 2,}

\item{legendTextSizeO}{Default =1,}

\item{legendTitleSizeI}{Default = 1,}

\item{legendTextSizeI}{Default =0.5,}

\item{animateOn}{Default = T,}

\item{delay}{Default = 100,}

\item{scenRef}{Default = NULL}

\item{extension}{Default =F,}

\item{boundaryRegShape}{Default = NULL,}

\item{boundaryRegShpFolder}{Default= NULL . Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default="")}

\item{boundaryRegShpFile}{Default=NULL . Suggested paste("ne_10m_admin_0_countries",sep  Default=""),}

\item{boundaryRegCol}{Default=NULL. Suggested "NAME_0",}

\item{boundaryRegionsSelect}{Default = NULL,}

\item{fillcolorNA}{Default = NULL}

\item{extendedFillColor}{Default ="grey75",}

\item{extendedBGColor}{Default ="lightblue1",}

\item{extendedHighLightColor}{Default ="cornsilk1",}

\item{extendedLabelsColor}{Default ="grey30",}

\item{extdendedLabelSize}{Default =0.7,}

\item{extendedShape}{Default =NULL,}

\item{extendedShapeCol}{Default =NULL,}

\item{expandPercent}{Default =2}

\item{projX}{Default = projX="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"}

\item{figWidth}{Default =9}

\item{figHeight}{Default =7}

\item{scaleRange}{Default NULL. Dataframe with columns param, maxScale, minScale to indicate maximum and minumum values for a parameter scale.}

\item{paramsSelect}{Default ="All"}

\item{indvScenarios}{Default =T,}

\item{GCMRCPSSPPol}{Default = F,}

\item{multiFacetCols}{Default ="scenarioRCP",}

\item{multiFacetRows}{Default ="scenarioGCM",}

\item{legendOutsideMulti}{Default = NULL,}

\item{legendPositionMulti}{Default = NULL,}

\item{legendTitleSizeMulti}{Default = NULL,}

\item{legendTextSizeAnim}{Default = NULL,}

\item{legendTextSizeMulti}{Default = NULL,}

\item{refGCM}{Default = NULL , eg. "gfdl-esm2m"}

\item{refRCP}{Default = NULL , eg. "rcp2p6"}

\item{chosenRefMeanYears}{Default=NULL}

\item{mapTitleSize}{Default=0.5}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the metis package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
