% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.mapsProcess.R
\name{metis.mapsProcess}
\alias{metis.mapsProcess}
\title{metis.mapsProcess}
\usage{
metis.mapsProcess(polygonDataTables = NULL, gridDataTables = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  folderName = "folderNameDefault", xRange = "All", labels = F,
  labelsSize = 1.2, subRegShape = NULL, subRegShpFolder = NULL,
  subRegShpFile = NULL, subRegCol = NULL, dirNameAppend = "",
  nameAppend = "", legendOutsideSingle = T,
  legendOutsidePosition = NULL, legendPosition = NULL,
  legendFixedBreaks = 5, legendTitleSizeO = 2, legendTextSizeO = 1,
  legendTitleSizeI = 1.5, legendTextSizeI = 1, animateOn = T,
  fps = 1, scenRef = NULL, extension = F, boundaryRegShape = NULL,
  boundaryRegShpFolder = NULL, boundaryRegShpFile = NULL,
  boundaryRegCol = NULL, boundaryRegionsSelect = NULL,
  extendedLabels = T, extendedFillColor = "grey75",
  extendedBGColor = "lightblue1", extendedHighLightColor = "cornsilk1",
  extendedLabelsColor = "grey30", extdendedLabelSize = 0.7,
  extendedShape = NULL, extendedShapeCol = NULL, expandPercent = 3,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  figWidth = 6, figHeight = 7, scaleRange = NULL,
  paramsSelect = "All", indvScenarios = "All", scensSelect = "All",
  GCMRCPSSPPol = F, multiFacetCols = "scenarioGCM",
  multiFacetRows = "scenarioRCP", legendOutsideMulti = T,
  legendPositionMulti = NULL, legendTitleSizeMulti = NULL,
  legendTextSizeAnim = NULL, legendTextSizeMulti = NULL,
  refGCM = NULL, refRCP = NULL, chosenRefMeanYears = NULL,
  mapTitleSize = 0.5, facetLabelSizeMulti = 3,
  facetLabelSizeGCMRCP = 1.5, numeric2Cat_list = NULL, diffOn = F,
  frameShow = T, pdfpng = "png", fillcolorNA = "gray",
  fillshowNA = NA, fillcolorNULL = "gray", legendSingleColorOn = F,
  legendSingleValue = 0, legendSingleColor = "white")
}
\arguments{
\item{polygonDataTables}{Default = NULL,}

\item{gridDataTables}{Default = NULL,}

\item{dirOutputs}{Default = paste(getwd(),"/outputs",sep=""),}

\item{folderName}{Default ="folderNameDefault",}

\item{xRange}{Default ="All",}

\item{labels}{Default = F,}

\item{labelsSize}{Default = 1.2,}

\item{subRegShape}{Default = NULL,}

\item{subRegShpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),}

\item{subRegShpFile}{Default = paste("gadm36_1",sep=""),}

\item{subRegCol}{Default ="NAME_1",}

\item{dirNameAppend}{Default =""}

\item{nameAppend}{Default =""}

\item{legendOutsideSingle}{Default =F, Single plots by default have legends inside. This can be moved out if wanted.}

\item{legendOutsidePosition}{Default = NULL, # "right","left","top","bottom", "center"}

\item{legendPosition}{Default = NULL, # c("RIGHT','top') - RIGHT LEFT TOP BOTTOM}

\item{legendFixedBreaks}{Default = "5",}

\item{legendTitleSizeO}{Default = 2,}

\item{legendTextSizeO}{Default =1,}

\item{legendTitleSizeI}{Default = 1,}

\item{legendTextSizeI}{Default =0.5,}

\item{animateOn}{Default = T,}

\item{fps}{Default = 1,}

\item{scenRef}{Default = NULL}

\item{extension}{Default =F,}

\item{boundaryRegShape}{Default = NULL,}

\item{boundaryRegShpFolder}{Default= NULL . Suggested paste(getwd(),"/dataFiles/gis/naturalEarth",sep  Default="")}

\item{boundaryRegShpFile}{Default=NULL . Suggested paste("ne_10m_admin_0_countries",sep  Default=""),}

\item{boundaryRegCol}{Default=NULL. Suggested "NAME_0",}

\item{boundaryRegionsSelect}{Default = NULL,}

\item{extendedLabels}{Default = T}

\item{extendedFillColor}{Default ="grey75",}

\item{extendedBGColor}{Default ="lightblue1",}

\item{extendedHighLightColor}{Default ="cornsilk1",}

\item{extendedLabelsColor}{Default ="grey30",}

\item{extdendedLabelSize}{Default =0.7,}

\item{extendedShape}{Default =NULL,}

\item{extendedShapeCol}{Default =NULL,}

\item{expandPercent}{Default =2}

\item{projX}{Default = projX="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"}

\item{figWidth}{Default =9}

\item{figHeight}{Default =7}

\item{scaleRange}{Default NULL. Dataframe with columns param, maxScale, minScale to indicate maximum and minumum values for a parameter scale.}

\item{paramsSelect}{Default ="All"}

\item{indvScenarios}{Default ="All", If only want to run single scenarios without comparing with others}

\item{scensSelect}{Default ="All", choose scenarios from data}

\item{GCMRCPSSPPol}{Default = F,}

\item{multiFacetCols}{Default ="scenarioRCP",}

\item{multiFacetRows}{Default ="scenarioGCM",}

\item{legendOutsideMulti}{Default = NULL,}

\item{legendPositionMulti}{Default = NULL,}

\item{legendTitleSizeMulti}{Default = NULL,}

\item{legendTextSizeAnim}{Default = NULL,}

\item{legendTextSizeMulti}{Default = NULL,}

\item{refGCM}{Default = NULL , eg. "gfdl-esm2m"}

\item{refRCP}{Default = NULL , eg. "rcp2p6"}

\item{chosenRefMeanYears}{Default=NULL}

\item{mapTitleSize}{Default=0.5}

\item{facetLabelSizeMulti}{Default =3}

\item{facetLabelSizeGCMRCP}{Default =1.5}

\item{numeric2Cat_list}{Default=NULL,}

\item{diffOn}{Default = F. Whether to calculate diff values between scenarios.}

\item{frameShow}{Default = T. Whether to plot frame around maps and facets.}

\item{pdfpng}{Save IO figures as pdf or png. Type=String. Options: 'pdf' or 'png'. Default = 'png'}

\item{fillcolorNA}{Default="gray",}

\item{fillshowNA}{Default=NA,}

\item{fillcolorNULL}{Default="gray"}

\item{legendSingleColorOn}{Default=F,}

\item{legendSingleValue}{Default =0,}

\item{legendSingleColor}{Default="white"}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the metis package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
