% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.prepGrid.R
\name{metis.prepGrid}
\alias{metis.prepGrid}
\title{metis.prepGrid}
\usage{
metis.prepGrid(demeterFolders = NULL, demeterScenarios = NULL,
  demeterTimesteps = seq(from = 2005, to = 2100, by = 5),
  demeterUnits = NULL, tethysFolders = NULL, tethysScenarios = NULL,
  tethysUnits = NULL, tethysFiles = c("wddom", "wdelec", "wdirr",
  "wdliv", "wdmfg", "wdmin", "wdnonag", "wdtotal"), xanthosFiles = NULL,
  xanthosScenarios = NULL, xanthosScenarioAssign = NULL,
  xanthosCoordinatesPath = NULL, xanthosGridAreaHecsPath = NULL,
  tethysFilesScarcity = NULL, xanthosFilesScarcity = NULL,
  spanLowess = 0.25, popFolder = NULL, popFiles = NULL,
  biaFolder = NULL, biaFiles = NULL, popUnits = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""), folderName = NULL,
  saveFormat = "rds", filterYears = NULL)
}
\arguments{
\item{demeterFolders}{Full path to demeter outputs}

\item{demeterScenarios}{Name of demeter scenario}

\item{demeterTimesteps}{Default is seq(from=2005,to=2100,by=5)}

\item{demeterUnits}{No Default}

\item{tethysFolders}{Folder for tethys results}

\item{tethysScenarios}{Scenario name for tethys run}

\item{tethysUnits}{No Default}

\item{tethysFiles}{Default =c("wddom","wdelec","wdirr","wdliv","wdmfg","wdmin","wdnonag","wdtotal"),}

\item{xanthosFiles}{Xanthos Files to Read}

\item{xanthosScenarios}{Default=NULL,}

\item{xanthosScenarioAssign}{Default NULL. Scenario name if testing single scenario.}

\item{xanthosCoordinatesPath}{paste(getwd(),"/dataFiles/grids/xanthosCoords/coordinates.csv",sep="")}

\item{xanthosGridAreaHecsPath}{=paste(getwd(),"/dataFiles/grids/xanthosRunsChris/reference/Grid_Areas_ID.csv",sep=""),}

\item{tethysFilesScarcity}{Default=NULL,}

\item{xanthosFilesScarcity}{Default=NULL,}

\item{spanLowess}{Default = 0.25}

\item{popFolder}{Default = <-paste(getwd(),"/dataFiles/grids/griddedIDsPop/",sep="")}

\item{popFiles}{Default = <-"grid_pop_map"}

\item{biaFolder}{Default = <-paste(getwd(),"/dataFiles/grids/griddedIDsbia/",sep="")}

\item{biaFiles}{Default = <-"grid_bia_map"}

\item{popUnits}{Default = <-"person"}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{folderName}{Default=NULL}

\item{saveFormat}{Default="rds". Choose between "rds" (Native R much faster) or "csv" or "both".}

\item{filterYears}{Default=seq(1980,2100,by=5)}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
