% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.prepGrid.R
\name{metis.prepGrid}
\alias{metis.prepGrid}
\title{metis.prepGrid}
\usage{
metis.prepGrid(demeterFolder = "NA", demeterScenario = "NA",
  demeterTimesteps = seq(from = 2005, to = 2100, by = 5),
  demeterUnits = "NA", tethysFolder = "NA", tethysScenario = "NA",
  tethysUnits = "NA", tethysFiles = c("wddom", "wdelec", "wdirr",
  "wdliv", "wdmfg", "wdmin", "wdnonag", "wdtotal"), xanthosFolder = "NA",
  xanthosFiles = "NA", xanthosScenarioAssign = "NA",
  xanthosCoordinatesPath = "NA", xanthosGridAreaHecsPath = "NA",
  scarcityXanthosRollMeanWindow = 10, spanLowess = 0.25,
  popFolder = "NA", popFiles = "NA", popUnits = "NA",
  dirOutputs = paste(getwd(), "/outputs", sep = ""), reReadData = 1,
  gridMetisData = paste(getwd(), "/outputs/Grids/gridMetis.RData", sep =
  ""), sqliteUSE = F, sqliteDBNamePath = paste(getwd(),
  "/outputs/Grids/gridMetis.sqlite", sep = ""))
}
\arguments{
\item{demeterFolder}{Full path to demeter outputs}

\item{demeterScenario}{Name of demeter scenario}

\item{demeterTimesteps}{Default is seq(from=2005,to=2100,by=5)}

\item{demeterUnits}{No Default}

\item{tethysFolder}{Folder for tethys results}

\item{tethysScenario}{Scenario name for tethys run}

\item{tethysUnits}{No Default}

\item{tethysFiles}{Default =c("wddom","wdelec","wdirr","wdliv","wdmfg","wdmin","wdnonag","wdtotal"),}

\item{xanthosFolder}{Xanthos Folder Path}

\item{xanthosFiles}{Xanthos Files to Read}

\item{xanthosScenarioAssign}{Default "NA". Scenario name if testing single scenario.}

\item{xanthosCoordinatesPath}{paste(getwd(),"/dataFiles/grids/xanthosCoords/coordinates.csv",sep="")}

\item{xanthosGridAreaHecsPath}{=paste(getwd(),"/dataFiles/grids/xanthosRunsChris/reference/Grid_Areas_ID.csv",sep=""),}

\item{scarcityXanthosRollMeanWindow}{Default = 10,}

\item{spanLowess}{Default = 0.25}

\item{popFolder}{Default = <-paste(getwd(),"/dataFiles/grids/griddedIDsPop/",sep="")}

\item{popFiles}{Default = <-"grid_pop_map"}

\item{popUnits}{Default = <-"person"}

\item{dirOutputs}{Default =paste(getwd(),"/outputs",sep=""),}

\item{reReadData}{Default =1,}

\item{gridMetisData}{Default = paste(dirOutputs, "/Grids/gridMetis.RData", sep = "")}

\item{sqliteUSE}{Default = T,}

\item{sqliteDBNamePath}{Default = paste(getwd(),"/outputs/Grids/gridMetis.sqlite", sep = "")}
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function prepares gridded data for use with other metis modules.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
