% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.R
\docType{package}
\name{metis}
\alias{metis}
\title{metis}
\description{
\itemize{
\item Github: https://github.com/JGCRI/metis
\item Webpage: https://jgcri.github.io/metis/
\item Cheatsheet: https://github.com/JGCRI/metis/blob/master/metisCheatsheet.pdf}
}
