% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.readgcam.R
\name{metis.readgcam}
\alias{metis.readgcam}
\title{metis.readgcam}
\usage{
metis.readgcam(gcamdatabasePath = NULL, gcamdatabaseName = NULL,
  queryxml = "metisQueries.xml", queryPath = paste(getwd(),
  "/dataFiles/gcam", sep = ""), scenOrigNames = NULL,
  scenNewNames = NULL, reReadData = T, dataProj = "dataProj.proj",
  dataProjPath = paste(getwd(), "/outputs", sep = ""),
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  regionsSelect = NULL, queriesSelect = "All", paramsSelect = "All",
  RawQueryCSV = NULL)
}
\arguments{
\item{gcamdatabasePath}{Path to gcam database folder}

\item{gcamdatabaseName}{Name of gcam database}

\item{queryxml}{Name of the query.xml file. By default it is "metisQueries.xml"}

\item{queryPath}{Folder that contains the query.xml file.By default it is
the same folder as specified by gcamdatabasePath}

\item{scenOrigNames}{Original Scenarios names in GCAM database in a string vector.
For example c('scenario1','scenario2).}

\item{scenNewNames}{New Names which may be shorter and more useful for figures etc.
Default will use Original Names. For example c('scenario1','scenario2)}

\item{reReadData}{If TRUE will read the GCAM data base and create a queryData.proj file
in the same folder as the GCAM database. If FALSE will load a '.proj' file if a file
with full path is provided otherwise it will search for a dataProj.proj file in the existing
folder which may have been created from an old run.}

\item{dataProj}{Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.}

\item{dataProjPath}{Folder that contains the dataProj or where it will be produced.
By default it is the same folder as specified by gcamdatabasePath}

\item{dirOutputs}{Full path to directory for outputs}

\item{regionsSelect}{The regions to analyze in a vector. Example c('Colombia','Argentina'). Full list:
c(USA, Africa_Eastern, Africa_Northern, Africa_Southern, Africa_Western, Australia_NZ, Brazil, Canada
Central America and Caribbean, Central Asia, China, EU-12, EU-15, Europe_Eastern, Europe_Non_EU,
European Free Trade Association, India, Indonesia, Japan, Mexico, Middle East, Pakistan, Russia,
South Africa, South America_Northern, South America_Southern, South Asia, South Korea, Southeast Asia,}

\item{queriesSelect}{Default = "All". Vector of queries to read from the queryxml for example
c("Total final energy by aggregate end-use sector", "Population by region"). The queries must be
availble in the queryxml file. Current list of queries and generated paramaters are:
\itemize{
\item "Total final energy by aggregate end-use sector". Parameters generated: energyFinalConsumBySecEJ.
\item "primary energy consumption by region (direct equivalent)".
Parameters generated: energyPrimaryByFuelEJ
\item "Electricity generation by aggregate technology". Parameters generated: elecByTechTWh
\item "water withdrawals by sector". Parameters generated: watWithdrawBySec
\item "water consumption by sector". Parameters generated: watConsumBySec
\item "water withdrawals by crop". Parameters generated: watWithdrawByCrop
\item "biophysical water demand by crop type and land region". Parameters generated: watBioPhysCons
\item "water withdrawals by water mapping source". Parameters generated: watIrrWithdrawBasin
\item "water consumption by water mapping source". Parameters generated: watIrrConsBasin
\item "GDP per capita MER by region". Where MER is "Market Exchange Rate".
Parameters generated: gdpPerCapita.
\item "GDP MER by region". Where MER is "Market Exchange Rate".
Parameters generated: gdp, gdpGrowthRate
\item "Population by region". Parameters generated: pop.
\item "ag production by tech". Where technologies signify irrigated or rainfed.
Parameters generated: agProdbyIrrRfd
\item "Ag Production by Crop Type". Parameters generated: agProdBiomass, agProdForest, agProdByCrop
\item "land allocation by crop and water source". Parameters generated: landIrrRfd
\item "aggregated land allocation". Parameters generated: landAlloc
\item "Land Use Change Emission". Parameters generated: emissLUCFut
\item "GHG emissions by subsector". Parameters generated: ghgEmissByGHGGROUPS, ghgEmissionByGHG
\item "CO2 emissions by sector". Parameters generated:emissCO2BySector
\item "nonCO2 emissions by sector". Parameters generated: emissCO2NonCO2BySectorGWPAR5, emissCO2NonCO2BySectorGTPAR5,emissNonCO2BySectorOrigUnits
}}

\item{paramsSelect}{Default = "All". If desired dplyr::select a subset of paramaters to analyze from the full list of parameters:
c(# Energy
"energyPrimaryByFuelEJ","energyPrimaryRefLiqProdEJ",
"energyFinalConsumBySecEJ","energyFinalByFuelBySectorEJ","energyFinalSubsecByFuelTranspEJ",
"energyFinalSubsecByFuelBuildEJ", "energyFinalSubsecByFuelIndusEJ","energyFinalSubsecBySectorBuildEJ",
"energyPrimaryByFuelMTOE","energyPrimaryRefLiqProdMTOE",
"energyFinalConsumBySecMTOE","energyFinalbyFuelMTOE","energyFinalSubsecByFuelTranspMTOE",
"energyFinalSubsecByFuelBuildMTOE", "energyFinalSubsecByFuelIndusMTOE","energyFinalSubsecBySectorBuildMTOE",
"energyPrimaryByFuelTWh","energyPrimaryRefLiqProdTWh",
"energyFinalConsumBySecTWh","energyFinalbyFuelTWh","energyFinalSubsecByFuelTranspTWh",
"energyFinalSubsecByFuelBuildTWh", "energyFinalSubsecByFuelIndusTWh","energyFinalSubsecBySectorBuildTWh",
# Electricity
"elecByTechTWh","elecCapByFuel","elecFinalBySecTWh","elecFinalByFuelTWh",
# Transport
"transportPassengerVMTByMode", "transportFreightVMTByMode", "transportPassengerVMTByFuel", "transportFreightVMTByFuel",
# Water
"watConsumBySec", "watWithdrawBySec", "watWithdrawByCrop", "watBioPhysCons", "watIrrWithdrawBasin","watIrrConsBasin",
# Socio-economics
"gdpPerCapita", "gdp", "gdpGrowthRate", "pop",
# Agriculture
"agProdbyIrrRfd","agProdBiomass", "agProdForest", "agProdByCrop",
# Land use
"landIrrRfd", "landAlloc","landAllocByCrop",
# Emissions
"emissLUC", "emissCO2BySector","emissCO2NonCO2BySectorGWPAR5","emissCO2NonCO2BySectorGTPAR5","emissNonCO2BySectorOrigUnits",
"emissNonCO2ByResProdGWPAR5", "emissTotalFFIBySec","emissMethaneBySource",
"emissCO2BySectorNonCO2GWPAR5", "emissCO2BySectorNonCO2GWPAR5LUC", "emissTotalBySec","emissCO2BySectorNoBio")}

\item{RawQueryCSV}{Optional.}
}
\value{
A list with the scenarios in the gcam database, queries in the queryxml file and a
tibble with gcam data formatted for metis charts.
}
\description{
This function connects to a gcamdatabase and uses a query file to
out results into a table ready for plotting.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
