% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis.templates.R
\name{metis.templates}
\alias{metis.templates}
\alias{metis.printPdfPng}
\alias{metis.chartsThemeLight}
\title{metis.templates}
\usage{
metis.printPdfPng(figure, dir, filename, figWidth = 13, figHeight = 9,
  pdfpng = "png")

metis.chartsThemeLight()
}
\arguments{
\item{figure}{Figure to be printed in function metis.printPdfPng}

\item{dir}{Directory to print figure to in function metis.printPdfPng}

\item{filename}{Filename for figure printed in function metis.printPdfPng}

\item{figWidth}{Figure Width in inches for figures to be printed in function metis.printPdfPng}

\item{figHeight}{Figure height in inches for figures to be printed in function metis.printPdfPng}

\item{pdfpng}{Either "pdf", "png" or "both" to define the format of output}
}
\value{
A list of different templates
}
\description{
This script holds various templates used for different scripts.
}
\details{
List of Templates in this script:
\itemize{
\item metis.printPdfPng: Function used to print charts to a pdf or png or both.
\item metis.chartsThemeLight: A light ggplot theme for charts
\item metis.tmapAnimate: A function to animate tmaps across a variable.
\item metis.tmapLayout: A fucntion to define tmap layouts
}
}
\keyword{charts,}
\keyword{maps,}
\keyword{print}
\keyword{templates,}
