% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.colors.R
\name{srn.colors}
\alias{srn.colors}
\title{srn.colors}
\usage{
srn.colors()
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for SRN modeling to the global environment
}
\examples{
List of Color Palettes
Once srn.colors() has been called all the following colors palettes are
added to the global environment. Each can be called to see the colors
and associated levels.
cbPalette
elec_tech_colors
elec_renew_colors
building_colors
trn_fuel_colors
enduse_fuel_numbered
enduse_colors
PAL_pri_ene
PAL_pri_fuelcost
PAL_emiss_sector
PAL_landuse
PAL_hydrogen
PAL_refliq
emiss_by_enduse_colors
biouse_colors
colorsX_Basic
colorsX_Gas
colorsX_Diff
colorsX_Diff5
colorsX_Absolute
colorsX_Absolute5
colorsX_Unassigned
colorsX_elec_subsec
colorsX_elec_finalNrgFuel
colorsX_elec_techs
colorsX_elec_sec
colorsX_finalNrg_sec
colorsX_PAL_pri_ene
colorsX_elec_tech_colors

EXample: Call srn.colors and build pie chart to view colors and labels
srn.colors()
pie(rep(1,length(colorsX_Basic)),label=names(colorsX_Basic),col=colorsX_Basic)
}
\keyword{colors,}
\keyword{palette}
