% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.colors.R
\name{srn.colors}
\alias{srn.colors}
\title{srn.colors}
\usage{
srn.colors()
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for SRN modeling
}
\examples{
srn.colors()
pie(rep(1,length(colorsX_Basic)),label=names(colorsX_Basic),col=colorsX_Basic)
}
\keyword{colors,}
\keyword{palette}
