% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.grid2poly.R
\name{srn.grid2poly}
\alias{srn.grid2poly}
\title{srn.grid2poly}
\usage{
srn.grid2poly()
}
\value{
A table with data by polygon ID for each shapefile provided
}
\description{
This function takes a .csv file with gridded lat, long data and aggregates
the data by spatial boundaries given different shapefiles.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
