% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.readgcam.R
\name{srn.readgcam}
\alias{srn.readgcam}
\title{srn.readgcam}
\usage{
srn.readgcam(gcamdatabasePath, gcamdatabaseName, queryxml, scenOrigNames,
  scenNewNames = NULL, reReadData = T, dataProj = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""))
}
\arguments{
\item{gcamdatabasePath}{Path to gcam database folder}

\item{gcamdatabaseName}{Name of gcam database}

\item{queryxml}{Full path to query.xml file}

\item{scenOrigNames}{Original Scenarios names in GCAM database in a string vector.
For example c("scenario1","scenario2).}

\item{scenNewNames}{New Names which may be shorter and more useful for figures etc.
Default will use Original Names. For example c("scenario1","scenario2)}

\item{reReadData}{If TRUE will read the GCAM data base and create a queryData.proj file
in the same folder as the GCAM database. If FALSE will load a ".proj" file if a file
with full path is provided otherwise it will search for a dataProj.proj file in the existing
folder which may have been created from an old run.}

\item{dirOutputs}{Full path to directory for outputs}
}
\value{
A list with the scenarios in the gcam database, queries in the queryxml file and a
tibble with gcam data formatted for srn charts.
}
\description{
This function connects to a gcamdatabase and uses a query file to
out results into a table ready for plotting.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
