% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srn.readgcam.R
\name{srn.readgcam}
\alias{srn.readgcam}
\title{srn.readgcam}
\usage{
srn.readgcam(gcamdatabasePath, gcamdatabaseName,
  queryxml = "srnQueries.xml", scenOrigNames, scenNewNames = NULL,
  reReadData = T, dataProj = "dataProj.proj",
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  regionsSelect = NULL, queriesSelect = "All")
}
\arguments{
\item{gcamdatabasePath}{Path to gcam database folder}

\item{gcamdatabaseName}{Name of gcam database}

\item{queryxml}{Full path to query.xml file}

\item{scenOrigNames}{Original Scenarios names in GCAM database in a string vector.
For example c('scenario1','scenario2).}

\item{scenNewNames}{New Names which may be shorter and more useful for figures etc.
Default will use Original Names. For example c('scenario1','scenario2)}

\item{reReadData}{If TRUE will read the GCAM data base and create a queryData.proj file
in the same folder as the GCAM database. If FALSE will load a '.proj' file if a file
with full path is provided otherwise it will search for a dataProj.proj file in the existing
folder which may have been created from an old run.}

\item{dataProj}{Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.}

\item{dirOutputs}{Full path to directory for outputs}

\item{regionsSelect}{The regions to analyze in a vector. Example c('Colombia','Argentina')}

\item{queriesSelect}{Default = "All". Vector of queries to read from the queryxml for example
c("Total final energy by aggregate end-use sector", "Population by region"). The queries must be
availble in the queryxml file. Current list of available paramaters are:
\itemize{
\item "Total final energy by aggregate end-use sector"
\item "GDP per capita MER by region" : Where MER is "Market Exchange Rate"
\item "GDP MER by region" : Where MER is "Market Exchange Rate"
\item "GDP Growth Rate (Percent)" : Calculated based on the GDP MER by region.
\item "Population by region"
\item "ag production by tech" : Where technologies signify irrigated or rainfed
}}
}
\value{
A list with the scenarios in the gcam database, queries in the queryxml file and a
tibble with gcam data formatted for srn charts.
}
\description{
This function connects to a gcamdatabase and uses a query file to
out results into a table ready for plotting.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
